==> DataProcessing/Notebooks/notebook00-merging files and adding columns.Rmd <==
---
title: "Merging Files"
output: html_notebook
---

Load in all of the files by setting the working directory and making a list of filenames to grab
Add the Date and View of the trail to seperate columns using the character string of the filename


```{r}
rm(list=ls(all=TRUE))
setwd("~/Post-doc/Data/Processed MatLab Files/Vertical")
path <- "C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Processed MatLab Files\\Vertical\\"
files <- list.files(path=path, pattern="*.csv")


ldf = lapply(files, function(x) {
      dat = read.table(x, header=T, sep=",")
      # Add column names
      names(dat) = c('X', 'Y', 'Z', 'Track')
      # Add a column with the Date of the trial
      dat$Date = substr(x,1,8)
      # Add a column with the View of the trial
      dat$View = substr(x,14,15)
      # Add a column with the filename
      dat$File.name = substr(x,1,15)
      return(dat)
})

library(dplyr)
df <- do.call("rbind", ldf)
## Date_view_track factor created
df$D_V_T <- paste(df$Date, df$View, df$Track, sep='_')
df$D_V_T <- as.factor(df$D_V_T)
## Date_view factor created
df$D_V <- paste(df$Date, df$View, sep='_') 
df$D_V <- as.factor(df$D_V)
df$Track <- as.factor(df$Track)
df$Date <- as.factor(df$Date)
df$View <- as.factor(df$View)
df$File.name <- as.factor(df$File.name)
levels(df$File.name)

head(df)
tail(df)
str(df)
```

Add columns of metadata based on the D_V of each row

```{r}
metadata <- read.csv("C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Metadata flat file 20230623.csv", header = T)
head(metadata)
tail(metadata)
metadata$File.name <- as.factor(metadata$File.name)
metadata$Date <- as.factor(metadata$Date)
metadata$Flow.rate <- as.factor(metadata$Flow.rate)
metadata$Chlorophyll <- as.factor(metadata$Chlorophyll)
metadata$Guano <- as.factor(metadata$Guano)
metadata$Light <- as.factor(metadata$Light)
metadata$Flume <- as.factor(metadata$Flume)
metadata$Flow.Direction <- as.factor(metadata$Flow.Direction)

str(metadata)

alldata <- merge(df, metadata)

head(alldata)
tail(alldata)
str(alldata)

delete.na <- function(DF, n=0) {
  DF[rowSums(is.na(DF)) <= n,]
}

TotalData <- delete.na(alldata, 2)
head(TotalData)
tail(TotalData)

##TotalData$Flow.rate <- as.factor(TotalData$Flow.rate)
##TotalData$Chlorophyll <- as.factor(TotalData$Chlorophyll)
##TotalData$Guano <- as.factor(TotalData$Guano)
##TotalData$Light <- as.factor(TotalData$Light)
##TotalData$Flume <- as.factor(TotalData$Flume)
##TotalData$Flow.Direction <- as.factor(TotalData$Flow.Direction)

str(TotalData)

```

Add horizontal and vertical heading angles, d and v values
```{r}
x1 <- TotalData$X[1:(nrow(TotalData)-1)]
x2 <- TotalData$X[2:nrow(TotalData)]
TotalData$dx[1:(nrow(TotalData)-1)] <- x2-x1

y1 <- TotalData$Y[1:(nrow(TotalData)-1)]
y2 <- TotalData$Y[2:nrow(TotalData)]
TotalData$dy[1:(nrow(TotalData)-1)] <- y2-y1

z1 <- TotalData$Z[1:(nrow(TotalData)-1)]
z2 <- TotalData$Z[2:nrow(TotalData)]
TotalData$dz[1:(nrow(TotalData)-1)] <- z2-z1

TotalData$d <- sqrt(TotalData$dx^2 + TotalData$dy^2 + TotalData$dz^2)

head(TotalData)
tail(TotalData)
```

```{r}
TotalData$vx <- TotalData$dx/(1/30)  ## velocity on the x axis
TotalData$vy <- TotalData$dy/(1/30)  ## velocity on the y axis
TotalData$vz <- TotalData$dz/(1/30)  ## velocity on the z axis
TotalData$v <- TotalData$d/(1/30) ## total velocity

TotalData$h.heading <- atan2(TotalData$dx, TotalData$dy)
TotalData$v.heading <- atan(TotalData$dz/(sqrt(TotalData$dx^2 + TotalData$dy^2)))

hist(TotalData$v.heading/(2*pi)*360)
hist(TotalData$h.heading/(2*pi)*360)

hist(TotalData$dz[abs(TotalData$dz)<.01],breaks = 100)
hist(TotalData$dy[abs(TotalData$dy)<.01],breaks = 100)
hist(TotalData$dx[abs(TotalData$dx)<.01],breaks = 100)

```

```{r}
save.image("~/Post-doc/Data/Vertical Data File (Nov 12 2024).RData")
```


```{r}







```
