==> DataProcessing/Notebooks/notebook01-distributions.Rmd <==
---
title: "Checking Distribution for Variables"
output: html_notebook
---

Plot and fit distributions of variables

```{r}

load("C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Total Merged Data File (Sep 20 2023).Rdata")

hist(log10(TotalData$v),
     xlab = "Velocity (Log^10 mm/s)")

```

Histogram of Velocity of all

Log distribution, 0's make up 3% of all data

```{r}
CC.TotalData <- na.omit(TotalData)
head(CC.TotalData)
10^mean(log10(CC.TotalData$v[CC.TotalData$v>0]))
sd(log10(CC.TotalData$v[CC.TotalData$v>0]))

levels(CC.TotalData$Date)
```
Now we're going to plot each unique krill and their bimodality with a dip test (less than 0.05 is multimodality)

```{r}
ind <- unique(CC.TotalData$D_V_T)
print(ind)
length(ind)
library(diptest)
library(DescTools)

tab <- matrix(data = NA, nrow = 429, ncol =5, byrow = T)
colnames(tab) <- c('dip.test', 'skew', 'mean.velocity', 'sd.velocity', 'Ind')

tab <- as.data.frame(tab)
tab$Ind <- ind
mean.vel <- NULL
sd.vel <- NULL
d.v <- NULL
s.v <- NULL

for (i in 1:length(ind)){
mean.v <- mean(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]])
sd.v <- sd(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]])

mean.vel <- rbind(mean.vel, mean.v)
sd.vel <- rbind(sd.vel, sd.v)


vels <- (CC.TotalData$v[CC.TotalData$D_V_T==ind[i]])
vels <- log10(vels[vels>0])
d <- dip.test(vels)
d.p <- d$p.value
d.v <- rbind(d.v, d.p)

s <- Skew(vels)
s.v <- rbind(s.v, s)
##}

##ind <- unique(CC.TotalData$D_V_T)

##for (i in 1:length(ind)){
 hist(log10(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]]),
     breaks = 50,
     xlab = "Velocity (Log^10 mm/s)",
     main = ind[i],
     sub = d.p)  ## change to d.p or s to print the dip test or skew value as the title instead
}

## to save each ind graph to working directory
for (i in 1:length(ind)){
  jpeg(filename=paste(ind[i], '.jpeg', sep = ''), width = 480, height = 480)
 hist(log10(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]]),
     breaks = 50,
     xlab = "Velocity (Log^10 mm/s)",
     main = ind[i],
     sub = d.p)
dev.off()
}

tab$skew <- s.v
tab$dip.test <- d.v
tab$mean.velocity <- (log10(mean.vel))
tab$sd.velocity <- (log10(sd.vel))
tab
write.table(tab, file = "~/Post-doc/Data/dip.test.skew.vels (July 24 2023).csv", sep = ",", col.names = TRUE)

plot(tab)


head(TotalData)
str(TotalData)
TotalData$Flow.rate <- as.factor(TotalData$Flow.rate)
TotalData$Chlorophyll <- as.factor(TotalData$Chlorophyll)

```


Now looking at turning angles

```{r}
TotalData$turn.anglexy <- atan2(TotalData$X, TotalData$Y)
TotalData$turn.angleyz <- atan2(TotalData$Y, TotalData$Z)
range(TotalData$turn.angle)


lth <- dim(TotalData)[1]
dx1 <- TotalData$dx[1:(lth-1)]
dx2 <- TotalData$dx[2:lth]
dy1 <- TotalData$dy[1:(lth-1)]
dy2 <- TotalData$dy[2:lth]
dz1 <- TotalData$dz[1:(lth-1)]
dz2 <- TotalData$dz[2:lth]
D <- (dx1*dx2)+(dy1*dy2)+(dz1*dz2)
d1 <- sqrt(dx1^2 + dy1^2 +dz1^2)
d2 <- sqrt(dx2^2 + dy2^2 +dz2^2)

dd <- D/d1/d2
hist(acos(dd)/pi*180)

TotalData$turn.angle <- c(NA, acos(D/d1/d2))/pi*180



head(TotalData)
CC.TotalData <- na.omit(TotalData)
head(CC.TotalData)
tail(CC.TotalData)
str(CC.TotalData)
CC.TotalData$Flow.rate <- as.factor(CC.TotalData$Flow.rate)
CC.TotalData$Chlorophyll<- as.factor(CC.TotalData$Chlorophyll)
CC.TotalData$Guano <- as.factor(CC.TotalData$Guano)
CC.TotalData$Light <- as.factor(CC.TotalData$Light)


CC.TotalData$Flow.rate <- as.numeric(CC.TotalData$Flow.rate)
CC.TotalData$Chlorophyll<- as.numeric(CC.TotalData$Chlorophyll)

head(CC.TotalData)

ind <- unique(CC.TotalData$D_V_T)

## to save ind histogram plots
for (i in 1:length(ind)){
jpeg(filename=paste(ind[i], '.jpeg', sep = ''), width = 480, height = 480)
hist(CC.TotalData$turn.angle[CC.TotalData$D_V_T==ind[i]],
     breaks = 50,
     xlab = "Turn Angles (degrees)",
     main = ind[i]) 
dev.off()
}

plot(TotalData$Flow.rate, TotalData$turn.angle, main = "", xlab = "Flow Rate (cm/s)", ylab = "Turn angle (degrees)")

plot(TotalData$Chlorophyll, TotalData$turn.angle, main = "", xlab = "Chlorophyll (mg/L)", ylab = "Turn angle (degrees)")

levels(CC.TotalData$Flow.rate)

levels(CC.TotalData$Flow.rate) <- c("No Flow", "Low Flow", "Medium Flow", "High Flow", "Extreme Flow")

levels(CC.TotalData$Chlorophyll)

levels(CC.TotalData$Chlorophyll) <- c("No Chlorophyll", "Low Chlorophyll",  "High Chlorophyll",  "High Chlorophyll",  "High Chlorophyll", "Low Chlorophyll", "Low Chlorophyll",  "Low Chlorophyll", "Low Chlorophyll", "Low Chlorophyll", "Low Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "Medium Chlorophyll", "Medium Chlorophyll",  "Medium Chlorophyll", "Medium Chlorophyll",  "Medium Chlorophyll", "Medium Chlorophyll", "Filament")
str(CC.TotalData)


library(ggplot2)
ggplot(CC.TotalData,aes(x=Flow.rate, y=log10(v), fill=Chlorophyll))+
  geom_boxplot(notch=F, notchwidth=0.3,outlier.shape=1,outlier.size=2, coef=1.5)+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=90,hjust=1,vjust=0.4))+
  theme(panel.grid.minor=element_blank())+
  labs(size= "",x = "Flow Rate (cm/s)", y = "Velocity (Log transformed)(mm/s)", title = "Light") +
  scale_fill_manual(values=c("White", "yellowgreen","green3", "darkgreen", "black"),name = "Chlorophyll (mg/L)",
                    labels=c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"))+
  facet_grid(~Light, scales = "free_x", space = "free")

ggplot(CC.TotalData,aes(x=Flow.rate, y=turn.angle, fill=Chlorophyll))+
  geom_boxplot(notch=F, notchwidth=0.3,outlier.shape=1,outlier.size=2, coef=1.5)+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=90,hjust=1,vjust=0.4))+
  theme(panel.grid.minor=element_blank())+
  labs(size= "",x = "Flow Rate (cm/s)", y = "Turn Angle (degrees)", title = "Light") +
  scale_fill_manual(values=c("White", "yellowgreen","green3", "darkgreen", "black"),name = "Chlorophyll (mg/L)",
                    labels=c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"))+
  facet_grid(~Light, scales = "free_x", space = "free")

```



Starting to look at bi-modality in the variables by merging the TotalData frame with the "tab" table
```{r}
##aggregating Complete cases of TotalData so we can merge it with tab data

## guano and light not working

AGG_TD <- aggregate(CC.TotalData, by = list(CC.TotalData$D_V_T, CC.TotalData$Guano, CC.TotalData$Light, CC.TotalData$Chlorophyll, CC.TotalData$Flow.rate), FUN = mean)
head(AGG_TD)
AGG_TD <- AGG_TD[ -c(6:7, 11:20)]
colnames(AGG_TD) <- c("Ind", "Guano", "Light","Chlorophyll", "Flow.rate", "X", "Y", "Z", "dx", "dy", "dz", "d", "vx", "vy", "vz", "v", "h.heading", "v.heading", "turn.anglexy", "turn.angleyz", "turn.angle")##, "vel.flow", "vel.turn.angle", "trim.X", "xsmooth", "ysmooth", "zsmooth", "log.vel.flow", "heading.pi", "pitch.perfect")
head(AGG_TD)
tail(AGG_TD)


tab_AGG <- merge(AGG_TD, tab, by = "Ind")
head(tab_AGG)
str(tab_AGG)
tab_AGG$Flow.Rate <- as.factor(tab_AGG$Flow.Rate)
tab_AGG$Chlorophyll <- as.factor(tab_AGG$Chlorophyll)

freq <- table(CC.TotalData$Light, CC.TotalData$Flow.rate, CC.TotalData$Guano)
print(freq)
prob <- prop.table(freq) ##Relative Frequency Table
print (prob)

##starting to plot the dip test and skew in the variables
plot(tab_AGG$Flow.Rate, tab_AGG$dip.test, xlab = "Flow Rate (cm/s)", ylab = "Dip Test (p.value)")
plot(tab_AGG$Flow.Rate, tab_AGG$skew, xlab = "Flow Rate (cm/s)", ylab = "Skew Test (p.value)")
plot(tab_AGG$Chlorophyll, tab_AGG$dip.test, xlab = "Chlorophyll (mg/L)", ylab = "Dip Test (p.value)")
plot(tab_AGG$Chlorophyll, tab_AGG$skew, xlab = "Chlorophyll (mg/L)", ylab = "Skew Test (p.value)")


## to save each ind graph
for (i in 1:length(ind)){
  jpeg(filename=paste(ind[i], '.jpeg', sep = ''), width = 480, height = 480)
plot(CC.TotalData$turn.angle[CC.TotalData$D_V_T==ind[i]], log10(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]]),
            xlab = "Turn Angles (degrees)",
     ylab = "Velocity (LOg^10, mm/s)",
          main = ind[i]) 
dev.off()
}
```


```{r}
hist(CC.TotalData$turn.angle, breaks = 100)



```



```{r}
save.image("~/Post-doc/Data/Total Merged Data File (Sep 20 2023).RData")

```
