==> DataProcessing/Notebooks/notebook02-velocity to edge effect.Rmd <==
---
title: "Velocity in relation to distance from edge"
output: html_notebook
---

Looking at relationship between velocity and the Y axis

```{r}
load("C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Total Data Merged File (Sep 20 2023).RData") 

library(ggplot2)
ind <- unique(CC.TotalData$D_V_T)
length(ind)

str(CC.TotalData)

## to save ind graphs of velocity by Y axis
for (i in 1:length(ind)){
jpeg(filename=paste(ind[i], '.jpeg', sep = ''), width = 480, height = 480)
plot(log10(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]]),
     CC.TotalData$Y[CC.TotalData$D_V_T==ind[i]],
     xlab = "Velocity (Log^10, mm/s)",
     ylab = "Y Axis (cm)",
     main = ind[i]) 
dev.off()
}

```


Looking at relationship between velocity and the X axis

```{r}
## to save individual graphs of X axis by velocity
for (i in 1:length(ind)){
  jpeg(filename=paste(ind[i], '.jpeg', sep = ''), width = 480, height = 480)
plot(log10(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]]),
     CC.TotalData$X[CC.TotalData$D_V_T==ind[i]],
     xlab = "Velocity (Log^10, mm/s)",
     ylab = "X Axis (cm)",
     main = ind[i])
dev.off()
}

```

Looking at relationship between velocity and the Z axis

```{r}
## to save individual graphs of z axis by velocity
for (i in 1:length(ind)){
  jpeg(filename=paste(ind[i], '.jpeg', sep = ''), width = 480, height = 480)
plot(log10(CC.TotalData$v[CC.TotalData$D_V_T==ind[i]]),
     CC.TotalData$Z[CC.TotalData$D_V_T==ind[i]],
     xlab = "Velocity (Log^10, mm/s)",
     ylab = "Z Axis (cm)",
     main = ind[i])
dev.off()
}

```

```{r}


```
