==> DataProcessing/Notebooks/notebook03-3D turn velocity.Rmd <==
---
title: "3D Turn velocity"
output: html_notebook
---

```{r}
rm(list=ls(all=TRUE))
load("C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Total Merged Data File (Sep 20 2023).RData")

head(TotalData) ## all Total Data rows
head(CC.TotalData) ##only complete cases of Total Data 

## Looking at vx, vy, vz, v and turn angles in xy, yz and total turn angles to calculate 3D turn velocities

## need to account for the fact that pivetting makes velocity different in the 3D rather than the 2 x 2D planes

## same equation as turn angle but for velocity??

## Looking at vx
##do we report angles? how is 3d vel diff from dx/dy/dz
## v vector accounting for flow (depends on tank location)
##acos(v1.v2/|v1||v2|) is angle
##if the krill were for example turning directly in front of where the flow was
##coming from how would that affect it
## maybe denote 2 angles

## between 180 and direction or change in direction????


##TotalData$turn.anglexy <- atan2(TotalData$X
##TotalData$turn.angleyz <- atan2(TotalData$Y


lth <- dim(TotalData)[1]

vx1 <- TotalData$vx[1:(lth-1)]
vy1 <- TotalData$vy[1:(lth-1)]
vz1 <- TotalData$vz[1:(lth-1)]


vx2 <- TotalData$vx[2:lth]
vy2 <- TotalData$vy[2:lth]
vz2 <- TotalData$vz[2:lth]

v1 <- sqrt((vx1^2)+(vy1^2)+(vz1^2))  ##velocity magnitude (one time step above)

v2 <- sqrt((vx2^2)+(vy2^2)+(vz2^2))
### velocity magnitude (1 time step below)

dotx <- (vx1*vx2)
doty <- (vy1*vy2)
dotz <- (vz1*vz2)

dotprod <- (dotx + doty + dotz)

dd <- dotprod/(v1*v2)

l <- 0
list2 <- ((acos(dd)/pi)*180)
list3 <- (v1+v2)/2
list4 <- c(l, list2)
list5 <- c(l, list3)
TotalData$vel.turn.angle <- list4  ##should be derivative of each other
TotalData$vel <- list5   

## can we get the derivative of the theta.angle

## save before doing vel.flow, just incase factor conversion doesn't work properly

save.image("~/Post-doc/Data/Vertical Data File (Nov 12 2024).RData")

```
