==> DataProcessing/Notebooks/notebook05-directional bearing.Rmd <==
---
title: "Directional Bearing"
output: html_notebook
---

```{r}

rm(list=ls(all=TRUE))
load("C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Total Merged Data File (March 6 2024).Rdata")

## Need to use turn angles and a point of reference (e.g. inlet valve) to determine directional bearing


library(circular)   ## loads the "circular" package

head(TotalData)
str(TotalData)
## We already have pitch and heading in TotalData
##Need to remember to convert heading into degrees - use:  TotalData$heading/(2*pi)*360

## head(CC.TotalData)  Flow and Chlorophyll are numeric - need to get back to factors with correct numbers
## str(CC.TotalData)
levels(FeedingDark$D_V)
FeedingDark <- subset(Feeding, Light=="Absent")
FeedingLight <- subset(Feeding, Light=="Present")
str(FeedingLight)
plot(droplevels(FeedingDark$Chlorophyll), FeedingDark$vel)


library(ggplot2)
ggplot(TotalData,aes(x=Flow.rate, y=h.heading/(2*pi)*360, fill=Chlorophyll))+
  geom_boxplot(notch=F, notchwidth=0.3,outlier.shape=1,outlier.size=2, coef=1.5)+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=90,hjust=1,vjust=0.4))+
  theme(panel.grid.minor=element_blank())+
  labs(size= "",x = "Flow Rate (cm/s)", y = "Horizontal Heading (degrees)", title = "Light") +
  scale_fill_manual(values=c("greenyellow", "yellowgreen","lightgreen", "green", "green3", "green4", "darkgreen"),name = "Chlorophyll (mg/L)",
                    labels=c("0", "4.3", "4.6", "6.1", "7.6", "13.5", "19"))+
  facet_grid(~Light, scales = "free_x", space = "free")
```


```{r}
library(circular)
TotalData$v.heading.pi <- TotalData$v.heading/(2*pi)*360


track.one<-TotalData[TotalData$Track=="1",]

track.two<-TotalData[TotalData$Track=="2",]

track.three<-TotalData[TotalData$Track=="3",]

track.four<-TotalData[TotalData$Track=="4",]

circ <- circular(track.one$turn.angleyz, type = c("angles"),
          units = c("degrees"),
          template = c("geographics"))

## S3 method for class 'circular'
## as(circ, control.circular=list(), ...) ## NOT WORKING CURRENTLY

## S3 method for class 'circular'
is(circ)

## S3 method for class 'circular'
print(circ, info=TRUE)


## S3 method for class 'circular'
plot(circLM, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 1.5,
bins = 24, ticks = TRUE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Light, Medium Flow, Horizontal Headings (degrees)", sub=NULL, xlab = "", ylab = "",
control.circle=circle.control())


##########################################

  ##ggplot2 approach

## redo with vertical headings, change range for turn angles?
write.table(FeedingDark, file = "~/Bigelow/Manuscripts/Feeding MS/FeedingDark.csv", sep = ",", col.names = TRUE)

head(AGG_Feed)
str(Feeding)
levels(FeedingDark$Chem) <- c("Ambient", "Guano", "Guano+CHL")
range(FeedingDark$turn.anglexysmooth)
FeedingLight$turn.anglexysmooth <- FeedingLight$turn.anglexysmooth/(2*pi)*360
f_mu <- mean(Feeding$turn.anglexysmooth, by=c(Chem, Flow.rate))

ggplot(mtcars, aes(disp, mpg)) +
  geom_point() +
  coord_radial(theta = 'x', start = 0, clip = 'on') ##,position =  position_dodge(0)

ggplot(df, aes(x = variable, y = value, fill = factor(dvmph))) +
  geom_col() +
  coord_polar()
##############################################################################
library(forcats)
library(ggplot2)
library(dplyr)
library(ggmulti)

#Code
data %>%
  filter(Chlorophyll %in% c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Extreme Chlorophyll", "Filament")) %>%
  ggplot(aes(x = heading.pi, fill = Chlorophyll))+
  geom_bar(stat = 'bin',  position = "fill", colour = "black", size = .25)+
  coord_polar(theta = "x", start = 3.142, direction=1)+
  theme_classic()+
  scale_x_continuous(limits = c(-179.999,180), ## if -180 labels overlap and look funny
                     breaks = seq(-180, 180, by = 45),
                     minor_breaks = seq(-180, 180, by = 15))+
  theme(text=element_text(size=20))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  theme(legend.key = element_rect(fill = "white"))+
  labs(x = "Horizontal Heading (degrees)", y = "Frequency of Heading", title = "") +
  scale_fill_manual(values=c("white", "gray90", "gray75","gray60","grey45", "grey30", "slategray"),name = "Chemical Cue",
                    labels=c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Extreme Chlorophyll", "Filament"))+
 facet_grid(facets = Light ~ Flow.Direction*Flume, labeller = label_both)

  ###########################################################
FeedingDark %>%
  filter(Chem %in% c("Ambient", "Guano", "Guano+CHL")) %>%
  ggplot(aes(x = turn.anglexysmooth, fill = Chem))+
  stat_bin_(data = FeedingDark,  position = "stack", binwidth = 15, colour = "black", size = .25)+
  coord_polar(theta = "x", start = 3.142, direction=1)+
  #geom_vline(data = Data2, aes(xintercept = ta), size = 1)+
  theme_classic()+
  scale_x_continuous(limits = c(-181,180), ## if -180 labels overlap and look funny
                     breaks = seq(-180, 180, by = 45),
                     minor_breaks = seq(-180, 180, by = 15))+
  theme(text=element_text(size=20))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  theme(legend.key = element_rect(fill = "white"))+
  labs(x = "Turn Angle (degrees)", y = "Frequency of Turn Angle", title = "") +
  scale_fill_manual(values=c("gray92", "gray", "gray35"),name = "Chemical Cue",
                    labels=c("Ambient", "Guano", "Guano+CHL"))+
 facet_grid(facets = Chem ~ Flow.rate, labeller = label_both)

#########################################################################

#ggplot(Feeding, aes(heading.pi-7.5, fill = Chem)) +
  geom_histogram(binwidth = 15, boundary = 0, colour = "black", size = .25) +
  #coord_radial(start = 0)+
    coord_polar(start = 3.142, direction=1) + ### adjusts position of 0 in plot
   ylim(-10, 25000)  +
  scale_x_continuous(limits = c(-179.99,180), ## if -180 labels overlap and look funny
                     breaks = seq(-180, 180, by = 45),
                     minor_breaks = seq(-180, 180, by = 15))+
  theme(text=element_text(size=20))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  theme(legend.key = element_rect(fill = "white"))+
  ##geom_vline(aes(heading.pi), xintercept = f_mu)+   ## adds mean angle line to plot
  labs(x = "Horizontal Heading (degrees)", y = "Proportional Frequency of Heading (%)", title = "") +
  scale_fill_manual(values=c("white", "gray", "gray35"),name = "Chemical Cue",
                    labels=c("Ambient", "Guano", "Guano+CHL"))+
 facet_grid(facets = Light ~ Flow.rate, labeller = label_both)

###############################################################################
####     coefficient of dispersion plot




CD <- read.csv(file = "~/Bigelow/Manuscripts/Feeding MS/CD.csv", sep = ',', header = TRUE)

CD %>%
  filter(Chem %in% c("Ambient", "Guano", "Guano+CHL")) %>%
  ggplot(aes(x = Data2$ta, y = 1))+
  coord_polar(theta = "x", start = 3.142, direction=1)+
  #geom_vline(data = Data2, aes(xintercept = ta), size = 1)+
  geom_segment(aes(x=0, y=0, xend=Data2$ta, yend=cd+.4), arrow = arrow(length=unit(.5, 'cm'), type = "open"))+
  theme_classic()+
  scale_x_continuous(limits = c(-181,180), ## if -180 labels overlap and look funny
                     breaks = seq(-180, 180, by = 45),
                     minor_breaks = seq(-180, 180, by = 15))+
  theme(text=element_text(size=20))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  labs(x = "Turn Angle (degrees)", y = "Coefficient of Dispersion (r)", title = "") +
  facet_grid(facets = Chem ~ Flow.rate, labeller = label_both)

########################


```


Plotting each subset of chlorophyll data in the XY plane

```{r}
circ0 <- circular(ChlA0$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ0, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Chlorophyll 0mg/L", sub="Control Conditions", xlab = "", ylab = "",
control.circle=circle.control())

##  jpeg(filename= "~/Post-doc/Graphs/Notebook 05 - directional bearing/Chla 0 turn.angle xy geographic.jpeg", units = "px", width = 480, height = 480)

circ4.3 <- circular(ChlA4.3$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ4.3, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = " Chlorophyll 4.3mg/L", sub="Chlorophyll Conditions", xlab = "", ylab = "",
control.circle=circle.control())

circ4.6 <- circular(ChlA4.6$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ4.6, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Chlorophyll 4.6mg/L", sub="Chlorophyll Conditions", xlab = "", ylab = "",
control.circle=circle.control())

circ6 <- circular(ChlA6$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ6, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Chlorophyll 6.1mg/L", sub="Chlorophyll Conditions", xlab = "", ylab = "",
control.circle=circle.control())

circ7 <- circular(ChlA7$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ7, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Chlorophyll 7.3mg/L", sub="Chlorophyll Conditions", xlab = "", ylab = "",
control.circle=circle.control())

circ13 <- circular(ChlA13$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ13, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Chorophyll 13.5mg/L", sub="Chlorophyll Conditions", xlab = "", ylab = "",
control.circle=circle.control())

circ19 <- circular(ChlA19$turn.anglexy, type = c("angles"),
          units = c("degrees"),
          template = c("geographic"))

plot(circ19, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = "Chlorophyll 19.5mg/L", sub="Chlorophyll Conditions", xlab = "", ylab = "",
control.circle=circle.control())

```

Do the same again but for Flow
```{r}

```

And for Light
```{r}

```

And for Guano
```{r}

```

Save the environment and remove unneeded dataframes, lists, etc
```{r}
#### want to look at absolute heading vs heading to flow. 
### Is this the same, eg., is 0 degrees (North) the inlet valve? (I think so but need to check this)



```




```{r}

## how do we account for krill headings while drifting backwards?

rm(ldf, s.v, sd.vel, D, d.p, d1, d2, dd, dotprod, dotx, doty, dotz, dx1, dx2, dy1, dy2, dz1, dz2, l, i, list2, list3, list4, list5, lth)
rm(mean.v, prob, s, sd.v, v1, v2, vels, vx1, vx2, vy1, vy2, vz1, vz2, x1, x2, y1, y2, z1, z2)

save.image("~/Post-doc/Data/Vertical Data File (Nov 12 2024).RData")

```
