==> DataProcessing/Notebooks/notebook08 - smoothing splines.Rmd <==
title: "Notebook08 - smoothing splines"
output: html_notebook
---

```{r}
rm(list=ls(all=TRUE))
load("C:\\Users\\Nicole Hellessey\\Documents\\Post-doc\\Data\\Total Merged Data File (Sep 20 2023).Rdata")
```

Remove ends of tank and clean data

```{r}
#### cut 1cm from each axis to get rid of edge effect
#### gives us only what's happening in center of tank 
### gives us defined segments to spline

## start by getting rid of front and back of tank 

hist(CC.TotalData$X)

#####################################################

range(CC.TotalData$X)
range(CC.TotalData$Y)
range(CC.TotalData$Z)
CC.TotalData$trim.X <- CC.TotalData$X

CC.TotalData$trim.X[which(CC.TotalData$trim.X < 0)] = CC.TotalData$trim.X+0.1
##CC.TotalData$X[which(CC.TotalData$X < 0)] = CC.TotalData$X+0.1 ##not sure if this would be better as it's not a mirror image flipping with the negative numbers

hist(CC.TotalData$trim.X)

```



```{r}

CC.TotalData$xsmooth <- runmed(x = CC.TotalData$trim.X, k = 31, na.action = 'na.omit')  ## optimum number for smoothing (try 31 to 101)
plot(CC.TotalData$X[1000:10000])
plot(CC.TotalData$xsmooth[1000:10000])

CC.TotalData$ysmooth <- runmed(x = CC.TotalData$Y, k = 31, na.action = 'na.omit')
plot(CC.TotalData$Y[1000:10000])
plot(CC.TotalData$ysmooth[1000:10000])


CC.TotalData$zsmooth <- runmed(x = CC.TotalData$Z, k = 31, na.action = 'na.omit')
plot(CC.TotalData$Z[1000:10000])
plot(CC.TotalData$zsmooth[1000:10000])


```

```{r}
##rm(ChlA0, ChlA4.3, ChlA4.6, ChlA5.5, ChlA6, ChlA7, ChlA13, ChlA19, f0, f.6, f3, f5, f8, l0, l1, g0, g1, kd1, kd2, kd3, kd4, kd5, kd6, kd7, kd8)
save.image("~/Post-doc/Data/Factor levels working correctly (Sep 20 2023).RData")
```



```{r}
smoothx1 <- CC.TotalData$xsmooth[1:(nrow(CC.TotalData)-1)]
smoothx2 <- CC.TotalData$xsmooth[2:nrow(CC.TotalData)]
CC.TotalData$smooth.dx[1:(nrow(CC.TotalData)-1)] <- smoothx2-smoothx1

smoothy1 <- CC.TotalData$ysmooth[1:(nrow(CC.TotalData)-1)]
smoothy2 <- CC.TotalData$ysmooth[2:nrow(CC.TotalData)]
CC.TotalData$smooth.dy[1:(nrow(CC.TotalData)-1)] <- smoothy2-smoothy1

smoothz1 <- CC.TotalData$zsmooth[1:(nrow(CC.TotalData)-1)]
smoothz2 <- CC.TotalData$zsmooth[2:nrow(CC.TotalData)]
CC.TotalData$smooth.dz[1:(nrow(CC.TotalData)-1)] <- smoothz2-smoothz1

CC.TotalData$smooth.d <- sqrt(CC.TotalData$smooth.dx^2 + CC.TotalData$smooth.dy^2 + CC.TotalData$smooth.dz^2)

head(CC.TotalData)
tail(CC.TotalData)
```



```{r}
CC.TotalData$smooth.vx <- CC.TotalData$smooth.dx/(1/30)  ## velocity on the x axis
CC.TotalData$smooth.vy <- CC.TotalData$smooth.dy/(1/30)  ## velocity on the y axis
CC.TotalData$smooth.vz <- CC.TotalData$smooth.dz/(1/30)  ## velocity on the z axis
CC.TotalData$smooth.v <- CC.TotalData$smooth.d/(1/30) ## total velocity

CC.TotalData$smooth.h.heading <- atan2(CC.TotalData$smooth.dx, CC.TotalData$smooth.dy)
CC.TotalData$heading.pi <- CC.TotalData$smooth.h.heading/(2*pi)*360
CC.TotalData$smooth.v.heading <- atan2(CC.TotalData$smooth.dz, (sqrt(CC.TotalData$smooth.dx^2 + CC.TotalData$smooth.dy^2)))
CC.TotalData$pitch.perfect <- CC.TotalData$smooth.v.heading/(2*pi)*360


hist(CC.TotalData$smooth.v.heading/(2*pi)*360)
hist(CC.TotalData$smooth.h.heading/(2*pi)*360)
hist(CC.TotalData$smooth.h.heading)

hist(CC.TotalData$smooth.dz[abs(CC.TotalData$smooth.dz)<.01],breaks = 100)
hist(CC.TotalData$smooth.dy[abs(CC.TotalData$smooth.dy)<.01],breaks = 100)
hist(CC.TotalData$smooth.dx[abs(CC.TotalData$smooth.dx)<.01],breaks = 100)

```

```{r}
CC.TotalData$turn.anglexysmooth <- atan2(CC.TotalData$xsmooth, CC.TotalData$ysmooth)
CC.TotalData$turn.angleyzsmooth <- atan2(CC.TotalData$ysmooth, CC.TotalData$zsmooth)


lth <- dim(CC.TotalData)[1]
dx1 <- CC.TotalData$smooth.dx[1:(lth-1)]
dx2 <- CC.TotalData$smooth.dx[2:lth]
dy1 <- CC.TotalData$smooth.dy[1:(lth-1)]
dy2 <- CC.TotalData$smooth.dy[2:lth]
dz1 <- CC.TotalData$smooth.dz[1:(lth-1)]
dz2 <- CC.TotalData$smooth.dz[2:lth]
D <- (dx1*dx2)+(dy1*dy2)+(dz1*dz2)
d1 <- sqrt(dx1^2 + dy1^2 +dz1^2)
d2 <- sqrt(dx2^2 + dy2^2 +dz2^2)

dd <- D/d1/d2

CC.TotalData$turn.angle.smooth <- c(NA, acos(dd))/pi*180
hist(CC.TotalData$turn.angle.smooth, breaks = 50)
head(CC.TotalData)
```


```{r}

lth <- dim(CC.TotalData)[1]

vx1 <- CC.TotalData$smooth.vx[1:(lth-1)]
vy1 <- CC.TotalData$smooth.vy[1:(lth-1)]
vz1 <- CC.TotalData$smooth.vz[1:(lth-1)]


vx2 <- CC.TotalData$smooth.vx[2:lth]
vy2 <- CC.TotalData$smooth.vy[2:lth]
vz2 <- CC.TotalData$smooth.vz[2:lth]

v1 <- sqrt((vx1^2)+(vy1^2)+(vz1^2))  ##velocity magnitude (one time step above)

v2 <- sqrt((vx2^2)+(vy2^2)+(vz2^2))
### velocity magnitude (1 time step below)

dotx <- (vx1*vx2)
doty <- (vy1*vy2)
dotz <- (vz1*vz2)

dotprod <- (dotx + doty + dotz)

dd <- dotprod/(v1*v2)

l <- 0
list2 <- ((acos(dd)/pi)*180)
list3 <- (v1+v2)/2
list4 <- c(l, list2)
list5 <- c(l, list3)
CC.TotalData$vel.turn.angle.smooth <- list4

head(CC.TotalData)

```


```{r}

str(CC.TotalData)
CC.TotalData$Flow.rate <- as.character(CC.TotalData$Flow.rate)
CC.TotalData$Flow.rate <- as.numeric(CC.TotalData$Flow.rate)

head(CC.TotalData) ## all TotalData rows
str(CC.TotalData)
CC.TotalData$vel.flow <- CC.TotalData$Flow.rate+CC.TotalData$vy
CC.TotalData$Flow.rate <- as.factor(CC.TotalData$Flow.rate)
str(CC.TotalData)
```
Don't forget to remake each factor level dataframe (E.g., CHlA0, f0, l0, g0, etc)
It's in Notebook 06

```{r}
rm(D, d1, d2, dd, dotprod, dotx, doty, dotz, dx1, dx2, dy1, dy2, dz1, dz2, list2, list3, list4, list5, v1, v2, vx1, vx2, vy1, vy2, vz1, vz2, x1, x2, y1, y2, z1, z2, l, lth, smoothx1, smoothx2, smoothy1, smoothy2, smoothz1, smoothz2)
save.image("~/Post-doc/Data/Total Merged Data File (Sep 20 2023).RData")
```
