==> DataProcessing/Notebooks/notebook09-statistical analysis.Rmd <==
---
title: "notebook09 - Statistical Analysis"
output: html_notebook
---

```{r}
rm(list=ls(all=TRUE))
load("~/Post-doc/Data/Vertical Flume Data Analaysis.RData")
###str(CC.TotalData)

AGG_Dark <- aggregate(FeedingDark, by = list(FeedingDark$D_V_T, FeedingDark$Chem, FeedingDark$Flow.rate), FUN = mean)
head(AGG_Dark)
AGG_Dark <- AGG_Dark[ -c(4:5, 9:18, 54:56)]
colnames(AGG_Dark) <- c("D_V_T", "Chem", "Flow.rate", "X", "Y", "Z", "dx", "dy", "dz", "d", "vx", "vy", "vz", "v", "h.heading", "v.heading", "turn.anglexy", "turn.angleyz", "turn.angle", "vel.turn.angle", "vel", "vel.flow", "trim.X", "xsmooth", "ysmooth", "zsmooth", "smooth.dx", "smooth.dy", "smooth.dz", "smooth.d", "smooth.vx", "smooth.vy", "smooth.vz", "smooth.v", "smooth.h.heading", "heading.pi", "smooth.v.heading", "pitch.perfect", "turn.anglexysmooth", "turn.angleyzsmooth", "turn.angle.smooth")
head(AGG_Dark)
tail(AGG_Dark)


AGG_Dark$Flow.Rate <- as.factor(AGG_Dark$Flow.Rate)
AGG_Dark$Chemical.Cue <- as.factor(AGG_Dark$Chemical.Cue)
```

Data for mean velocity, angular velocity, turn and heading angles each will be analyzed by a mixed model ANCOVA with chemical concentration, and light as fixed (treatment) effects, flow velocity as the co-variate and replicate as a random (blocking affect) with individual paths as the unit of analysis.

```{r}

##To run ANCOVA in R load the following packages:
library(car)
library(compute.es)
library(effects)
library(ggplot2)
library(multcomp)
library(pastecs)
library(WRS2)

Feeding <- rbind(LM, LH, DM, DH)
str(AGG_Feed)
Feeding$Chem <- as.factor(Feeding$Chem)
levels(FeedingDark$Chem)
##If you are using lm or aov make sure that you set the contrasts using the "contrasts" function before doing either aov or lm. R uses non-orthogonal contrasts by default which can mess everything up in an ANCOVA. If you want to set orthogonal contrasts use:

contrasts(dataname$factorvariable)=contr.poly()# of levels, i.e. 3) ##example
contrasts(Feeding$Chem)=contr.poly(3)
##then run your model as

head(AGG_Dark)

model.1=aov(dv~covariate+factorvariable, data=dataname) ## example
mod1=aov(smooth.vel ~ Flow.Rate*Chemical.Cue, data=AGG_Dark)
mod2 <- update(mod1, ~.-Chemical.Cue:Flow.Rate)
##To view the model use:

Anova(model.1, type="III") ## example
Anova(mod1, type="III")
Anova(mod2, type="III")
##Make sure you use capital "A" Anova here and not anova. This will give results using type III SS.

summary.lm(mod1) ##will give another summary and includes the R-sq. output. Example
summary.lm(mod2)

posth=glht(model.1, linfct=mcp(factorvariable="Tukey"))  ##gives the post-hoc Turkey analysis. Example
posth=glht(mod1, linfct=mcp(Chemical.Cue*Flow.Rate ="Tukey"))
summary(posth) ##shows the output in a nice format.

TukeyHSD(mod1, conf.level=.95) ## gives full Tukey post-hoc for all pairwise combinations

##If you want to test for homogeneity of regression slopes you can also include an interaction term for the IV and covariate. That would be:

model=aov(dv~covariate+IV+covariate:IV, data=dataname) ## example

mod2=aov(vel.flow~Flow.rate+Chem+Flow.rate:Chem, data=AGG_Feed) ## not right?
##If the interaction term is significant then you do not have homogeneity.

##################################

##linear mixed effect models
fit <- aov(v ~ Chem*Light*Flow.rate, data=AGG_Feed)
fit2 <- update(fit, ~.-Chem:Flow.rate:Light)# same thing minus 3-way interaction

##par(mfrow=c(2,2), mar=c(2,2,2,2), oma=c(2,2,2,2)) # optional layout 
##plot(fit) # diagnostic plots

summary(fit2) # display Type I ANOVA table
drop1(mod1,~.,test="F") # type III SS and F Tests

# Tukey Honestly Significant Differences
TukeyHSD(mod1) 

######################################################################

str(AGG_Dark)
AGG_Dark$D_V_T <- as.factor(AGG_Dark$D_V_T)
AGG_Dark$Chem <- as.factor(AGG_Dark$Chem)
AGG_Dark$Flow.rate <- as.factor(AGG_Dark$Flow.rate)
levels(AGG_Dark$Flow.rate)
AGG_Dark$Flow.rate <- relevel(AGG_Dark$Flow.rate, "Medium Flow")

## boxplots for velocity for each level of flow
library(ggplot2)
ggplot(AGG_Dark,aes(x=Flow.Rate, y=log10(smooth.vel), fill=Chemical.Cue))+
  geom_boxplot(notch=F, notchwidth=0.3,outlier.shape=1,outlier.size=2, coef=1.5)+
  stat_summary(geom = "errorbar", fun.min = mean, fun = mean, fun.max = mean, width = .75, linetype = "dashed", position = position_dodge())+
  theme_classic()+
  theme(axis.text=element_text(color="black"))+
  theme(text=element_text(size=20))+
  theme(axis.text.x=element_text(angle=90,hjust=1,vjust=0.4))+
  theme(panel.grid.minor=element_blank())+
  labs(size= "",x = "Flow Rate (cm s-1)", y = "Ground Velocity (logged, cm s-1)", title = "") +
  scale_fill_manual(values=c("white", "gray", "slategray"),name = "Chemical Cue",
  labels=c("Ambient", "Guano", "Guano+CHL"))
  

```

Stats on each level

```{r}
##Looking at other stats we could run
library(nlme)
library(lme4)
library(effects)


## example
d <- data.frame(
    Y = rnorm(48),  ### e.g. velocity
    subject = factor(rep(1:12, 4)),  ### e.g. D_V_T
    A = factor(rep(1:2, each=24)),   #### e.g. Light
    B = factor(rep(rep(1:2, each=12), 2)))   ### e.g. Flow.rate

summary(aov(Y ~ A*B + Error(subject/(A*B)), data=d))  # Standard repeated measures ANOVA

mode1 <- lme(Y~A*B, random = ~1|subject, data=d)  ## nlme

lmer(Y ~ A*B + (1|subject) + (1|A:subject) + (1|B:subject), data=d)  ### lme4



#########    our data using nlme package     ######### ## block trial, not track

mode1 <- lme(pitch.perfect~Chem*Light*Flow.rate, random= ~1|D_V_T, data=Feeding)

mode2 <- lme(pitch.perfect~Light*Flow.rate, random= ~1|D_V_T, data=Feeding)

anova(mode1, mode2) #if it shows only minor improvement, no need to include missing factor

anova(mode2)


#########    our data using l4me package     #########

model1 <- lmer(vel.flow ~ Flow.rate*Chem*Light + (1|D_V_T) + (1|Flow.rate:D_V_T) + (1|Chem:D_V_T) +  (1|Light:D_V_T), data=Feeding)  ###

model2 <- lmer(vel.flow ~ Flow.rate*Light + (1|D_V_T) + (1|Flow.rate:D_V_T) +  (1|Light:D_V_T), data=Feeding)

anova(model1, model2) #if it shows only minor improvement, no need to include missing factor

anova(model1)

######### plotting effect of the models    ##########
plot(effect("vel.flow*Flow.rate",se=TRUE, confidence.level=.95, mode2))


###### interaction plot with Lattice extra #########
library(latticeExtra)
xyplot(Feeding$vel.flow~Feeding$turn.angle|Feeding$Light*Feeding$Flow.rate,  main="Scatterplots by Light and Flow Rate", 
   ylab="Vel.flow", xlab="Turn Angle") +
   layer(panel.ablineq(lm(y ~ x), r.sq = TRUE,
   at = 0.75, adj=1:5), style = 4)


############## interaction plots ###########################

interaction.plot(Feeding$vel.flow, ## x axis
                 Feeding$Flow.rate, ## factor for lines
                 Feeding$turn.angle,  ## y axis
                 fun = mean,
                 type = c("b"), legend = TRUE,
                 trace.label = "Flow Rate",
                 ylim = c(-1,180),
                 main = "Lights On",
                 xlab = "vel.flow",
                 ylab = "turn angle",
                 col = CC.TotalData$Flow.rate, pch = c(1:9),
                 axes = TRUE)
#library(tidyr)
#library(dplyr)
#CC.TotalData %>% mutate_all(~replace(., is.nan(.), 0))
Feeding[is.nan(Feeding)] <- 0
range(Feeding$log.vel.flow)
##########################################################

rm(mode1)

rm(mode2)


mod1 <- lm(vel.flow~Flow.rate*Chlorophyll, data = Feeding)
summary(mode2)

```

#################################################################################

                 Kernel Density Plots

##################################################################################

Turn Angles first...

```{r}
rm(kd1, kd2, kd3, kd4, kd5, kd6, kd7, kd8)

### Kernel density plots ################### Turn Angles
par(mar= c(4,4,2,1))
par(mfrow=c(2,2))

################################ Flow Rate ########################
str(CC.TotalData)
levels(CC.TotalData$Flow.rate)

kd1 <- density(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="0"], na.rm= T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.030), col='blue', lwd=2, main = "Turn Angles Vs Flow Rate")

mean(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="0"], na.rm = T)
# Create the function.
getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}
getmode(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="0"])
median(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="0"], na.rm = T)


kd2 <- density(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="0.6"], na.rm = T)
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="3"], na.rm = T)
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="5.9"], na.rm = T)
lines(kd4, col='orange', lwd=2)

kd5 <- density(CC.TotalData$turn.angle[CC.TotalData$Flow.rate=="8.9"], na.rm = T)
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)

################################ Chlorophyll ########################
levels(CC.TotalData$Chlorophyll)

kd1 <- density(CC.TotalData$turn.angle[CC.TotalData$Chlorophyll=="No Chlorophyll"], na.rm= T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.030), col='blue', lwd=2, main = "Turn Angles Vs Chlorophyll")

kd2 <- density(CC.TotalData$turn.angle[CC.TotalData$Chlorophyll=="Low Chlorophyll"], na.rm= T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(CC.TotalData$turn.angle[CC.TotalData$Chlorophyll=="Medium Chlorophyll"], na.rm= T)
lines(kd3, col='green', lwd=2)

kd4 <- density(CC.TotalData$turn.angle[CC.TotalData$Chlorophyll=="High Chlorophyll"], na.rm= T)
lines(kd4, col='darkgreen', lwd=2)

kd5 <- density(CC.TotalData$turn.angle[CC.TotalData$Chlorophyll=="Filament"], na.rm= T)
lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

################################ Guano ########################
levels(CC.TotalData$Guano)

kd1 <- density(CC.TotalData$turn.angle[CC.TotalData$Guano=="Present"], na.rm = T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.030), col='red', lwd=2, main = "Turn Angles Vs Guano")

kd2 <- density(CC.TotalData$turn.angle[CC.TotalData$Guano=="Absent"], na.rm = T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

################################ Light ########################
levels(CC.TotalData$Light)

kd1 <- density(CC.TotalData$turn.angle[CC.TotalData$Light=="Present"], na.rm = T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.030), col='yellow', lwd=2, main = "Turn Angles Vs Light")

kd2 <- density(CC.TotalData$turn.angle[CC.TotalData$Light=="Absent"], na.rm = T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

```

Velocity (in relation to flow)...

```{r}
### Kernel density plots ################### Swimming velocity
##range(CC.TotalData$vel.flow)
## plot(log.vel.flow)
##log.vel.flow <- log10(CC.TotalData$vel.flow[CC.TotalData$vel.flow > 0])
#str(CC.TotalData)
#  vels <- (CC.TotalData$v[CC.TotalData$D_V_T==ind[i]])  ## vels <- log10(vels[vels>0])  ## from notebook01, could be useful to try this instead
################################ Flow Rate ########################
levels(CC.TotalData$Flow.rate)

####################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(2,2))


kd1 <- density(CC.TotalData$vel.flow[CC.TotalData$Flow.rate=="0"], na.rm = T)
plot(kd1, xlim = c(-15,22), ylim = c(0, 170), col='blue', lwd=2, main = "Velocity (in relation to flow, not logged) Vs Flow Rate")

kd2 <- density(CC.TotalData$vel.flow[CC.TotalData$Flow.rate=="0.6"], na.rm = T)
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CC.TotalData$vel.flow[CC.TotalData$Flow.rate=="3"], na.rm = T)
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CC.TotalData$vel.flow[CC.TotalData$Flow.rate=="5.9"], na.rm = T)
lines(kd4, col='orange', lwd=2)

kd5 <- density(CC.TotalData$vel.flow[CC.TotalData$Flow.rate=="8.9"], na.rm = T)
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)

################################ Chlorophyll ########################
levels(CC.TotalData$Chlorophyll)

kd1 <- density(CC.TotalData$vel.flow[CC.TotalData$Chlorophyll=="No Chlorophyll"], na.rm = T)
plot(kd1, xlim = c(-15,22), ylim = c(0, 80), col='blue', lwd=2, main = "Velocity (in relation to flow, not logged) Vs Chlorophyll")

kd2 <- density(CC.TotalData$vel.flow[CC.TotalData$Chlorophyll=="Low Chlorophyll"], na.rm = T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(CC.TotalData$vel.flow[CC.TotalData$Chlorophyll=="Medium Chlorophyll"], na.rm = T)
lines(kd3, col='green', lwd=2)

kd4 <- density(CC.TotalData$vel.flow[CC.TotalData$Chlorophyll=="High Chlorophyll"], na.rm = T)
lines(kd4, col='darkgreen', lwd=2)

kd5 <- density(CC.TotalData$vel.flow[CC.TotalData$Chlorophyll=="Filament"], na.rm = T)
lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

################################ Guano ########################
levels(CC.TotalData$Guano)

kd1 <- density(CC.TotalData$vel.flow[CC.TotalData$Guano=="Present"], na.rm = T)
plot(kd1, xlim = c(-14,21), ylim = c(0, 2), col='red', lwd=2, main = "Velocity (in relation to flow, not logged) Vs Guano")

kd2 <- density(CC.TotalData$vel.flow[CC.TotalData$Guano=="Absent"], na.rm = T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

################################ Light ########################
levels(CC.TotalData$Light)

kd1 <- density(CC.TotalData$vel.flow[CC.TotalData$Light=="Present"], na.rm = T)
plot(kd1, xlim = c(-14,21), ylim = c(0, 2), col='yellow', lwd=2, main = "Velocity (in relation to flow, not logged) Vs Light")

kd2 <- density(CC.TotalData$vel.flow[CC.TotalData$Light=="Absent"], na.rm = T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

```

Horizontal Headings...

```{r}
## use CC.TotalDat$heading.pi

rm(kd1, kd2, kd3, kd4, kd5)


### Kernel density plots ################### Headings
par(mar= c(4,4,2,1))
par(mfrow=c(2,2))

################################ Flow Rate ########################
levels(CC.TotalData$Flow.rate)

kd1 <- density(CC.TotalData$heading.pi[CC.TotalData$Flow.rate=="0"], na.rm=T)
plot(kd1, xlim = c(-180,180), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Flow Rate")

kd2 <- density(CC.TotalData$heading.pi[CC.TotalData$Flow.rate=="0.6"], na.rm=T)
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CC.TotalData$heading.pi[CC.TotalData$Flow.rate=="3"], na.rm=T)
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CC.TotalData$heading.pi[CC.TotalData$Flow.rate=="5.9"], na.rm=T)
lines(kd4, col='orange', lwd=2)

kd5 <- density(CC.TotalData$heading.pi[CC.TotalData$Flow.rate=="8.9"], na.rm=T)
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)

################################ Chlorophyll ########################
levels(CC.TotalData$Chlorophyll)

kd1 <- density(CC.TotalData$heading.pi[CC.TotalData$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(-180,180), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Chlorophyll")

kd2 <- density(CC.TotalData$heading.pi[CC.TotalData$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(CC.TotalData$heading.pi[CC.TotalData$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

kd4 <- density(CC.TotalData$heading.pi[CC.TotalData$Chlorophyll=="High Chlorophyll"], na.rm=T)
lines(kd4, col='darkgreen', lwd=2)

kd5 <- density(CC.TotalData$heading.pi[CC.TotalData$Chlorophyll=="Filament"], na.rm=T)
lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

################################ Guano ########################
levels(CC.TotalData$Guano)

kd1 <- density(CC.TotalData$heading.pi[CC.TotalData$Guano=="Present"], na.rm=T)
plot(kd1,  xlim = c(-180,180), ylim = c(0, 0.03), col='red', lwd=2, main = "Heading Vs Guano")

kd2 <- density(CC.TotalData$heading.pi[CC.TotalData$Guano=="Absent"], na.rm=T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

################################ Light ########################
levels(CC.TotalData$Light)

kd1 <- density(CC.TotalData$heading.pi[CC.TotalData$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(-180,180), ylim = c(0, 0.04), col='yellow', lwd=2, main = "Heading Vs Light")

kd2 <- density(CC.TotalData$heading.pi[CC.TotalData$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

```

Vertical Headings...

```{r}
##Use CC.TotalData$pitch.perfect
head(CC.TotalData)

### Kernel density plots ################### Pitch
par(mar= c(4,4,2,1))
par(mfrow=c(2,2))

################################ Flow Rate ########################
levels(CC.TotalData$Flow.rate)

kd1 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Flow.rate=="0"], na.rm=T)
plot(kd1, xlim = c(-95,95), ylim = c(0, 0.7), col='blue', lwd=2, main = "Pitch Vs Flow Rate")

kd2 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Flow.rate=="0.6"], na.rm=T)
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Flow.rate=="3"], na.rm=T)
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Flow.rate=="5.9"], na.rm=T)
lines(kd4, col='orange', lwd=2)

kd5 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Flow.rate=="8.9"], na.rm=T)
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)

################################ Chlorophyll ########################
levels(CC.TotalData$Chlorophyll)

kd1 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(-95,95), ylim = c(0, 0.45), col='blue', lwd=2, main = "Pitch Vs Chlorophyll")

kd2 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

kd4 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Chlorophyll=="High Chlorophyll"], na.rm=T)
lines(kd4, col='darkgreen', lwd=2)

kd5 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Chlorophyll=="Filament"], na.rm=T)
lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

################################ Guano ########################
levels(CC.TotalData$Guano)

kd1 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Guano=="Present"], na.rm=T)
plot(kd1, xlim = c(-95,95), ylim = c(0, 0.55), col='red', lwd=2, main = "Pitch Vs Guano")

kd2 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Guano=="Absent"], na.rm=T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

################################ Light ########################
levels(CC.TotalData$Light)

kd1 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(-95,95), ylim = c(0, 2.2), col='yellow', lwd=2, main = "Pitch Vs Light")

kd2 <- density(CC.TotalData$pitch.perfect[CC.TotalData$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

```

Turn Angles for each level of Flow

```{r}
##For flow = 0
f0 <- CC.TotalData[CC.TotalData$Flow.rate=="0",]
f.6 <- CC.TotalData[CC.TotalData$Flow.rate=="0.6",]
f3 <- CC.TotalData[CC.TotalData$Flow.rate=="3",]
f5.9 <- CC.TotalData[CC.TotalData$Flow.rate=="5.9",]
f8.9 <- CC.TotalData[CC.TotalData$Flow.rate=="8.9",]

g0 <- CC.TotalData[CC.TotalData$Guano=="Absent",]
g1 <- CC.TotalData[CC.TotalData$Guano=="Present",]

l0 <- CC.TotalData[CC.TotalData$Light=="Absent",]
l1 <- CC.TotalData[CC.TotalData$Light=="Present",]

CHL0 <- CC.TotalData[CC.TotalData$Chlorophyll=="No Chlorophyll",]
CHLL <- CC.TotalData[CC.TotalData$Chlorophyll=="Low Chlorophyll",]
CHLM <- CC.TotalData[CC.TotalData$Chlorophyll=="Medium Chlorophyll",]
CHLH <- CC.TotalData[CC.TotalData$Chlorophyll=="High Chlorophyll",]
CHLF <- CC.TotalData[CC.TotalData$Chlorophyll=="Filament",]


### Kernel density plots ################### Turn Angles
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

################################ Chlorophyll ########################
levels(f0$Chlorophyll)

##Flow = 0

kd1 <- density(f0$turn.angle[f0$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angles Vs Chlorophyll, Flow = 0")

kd2 <- density(f0$turn.angle[f0$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f0$turn.angle[f0$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

kd4 <- density(f0$turn.angle[f0$Chlorophyll=="High Chlorophyll"], na.rm=T)
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$turn.angle[f0$Chlorophyll=="Filament"], na.rm=T)
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

##Flow = 0.6


kd1 <- density(f.6$turn.angle[f.6$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angles Vs Chlorophyll, Flow = 0.6")

kd2 <- density(f.6$turn.angle[f.6$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f.6$turn.angle[f.6$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

kd4 <- density(f.6$turn.angle[f.6$Chlorophyll=="High Chlorophyll"], na.rm=T)
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$turn.angle[f0$Chlorophyll=="Filament"], na.rm=T)
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)


##Flow = 3

kd1 <- density(f3$turn.angle[f3$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angles Vs Chlorophyll, Flow = 3")

kd2 <- density(f3$turn.angle[f3$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f3$turn.angle[f3$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

kd4 <- density(f3$turn.angle[f3$Chlorophyll=="High Chlorophyll"], na.rm=T)
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$turn.angle[f0$Chlorophyll=="Filament"], na.rm=T)
##lines(kd5, col='black', lwd=2)


legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)


##Flow = 5.9

kd1 <- density(f5.9$turn.angle[f5.9$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angles Vs Chlorophyll, Flow = 5.9")

kd2 <- density(f5.9$turn.angle[f5.9$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f5.9$turn.angle[f5.9$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

##kd4 <- density(f5.9$turn.angle[f5.9$Chlorophyll=="High Chlorophyll"], na.rm=T)
##lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$turn.angle[f0$Chlorophyll=="Filament"], na.rm=T)
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

##Flow = 8.9

kd1 <- density(f8.9$turn.angle[f8.9$Chlorophyll=="No Chlorophyll"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angles Vs Chlorophyll, Flow = 8.9")

kd2 <- density(f8.9$turn.angle[f8.9$Chlorophyll=="Low Chlorophyll"], na.rm=T)
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f8.9$turn.angle[f8.9$Chlorophyll=="Medium Chlorophyll"], na.rm=T)
lines(kd3, col='green', lwd=2)

##kd4 <- density(f8.9$turn.angle[f8.9$Chlorophyll=="High Chlorophyll"], na.rm=T)
##lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$turn.angle[f0$Chlorophyll=="Filament"], na.rm=T)
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)

#####################################################################

################################# Guano ##################################

######################################################################

par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

##Flow = 0

kd1 <- density(f0$turn.angle[f0$Guano=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='red', lwd=2, main = "Turn Angles Vs Guano, Flow = 0")

kd2 <- density(f0$turn.angle[f0$Guano=="Absent"], na.rm=T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

##Flow = 0.6

kd1 <- density(f.6$turn.angle[f.6$Guano=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='red', lwd=2, main = "Turn Angles Vs Guano, Flow = 0.6")

kd2 <- density(f.6$turn.angle[f.6$Guano=="Absent"], na.rm=T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

##Flow = 3

kd1 <- density(f3$turn.angle[f3$Guano=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='red', lwd=2, main = "Turn Angles Vs Guano, Flow = 3")

kd2 <- density(f3$turn.angle[f3$Guano=="Absent"], na.rm=T)
lines(kd2, col='blue', lwd=2)

legend("topright",                              # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

##Flow = 5.9

kd1 <- density(f5.9$turn.angle[f5.9$Guano=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='red', lwd=2, main = "Turn Angles Vs Guano, Flow = 5.9")

kd2 <- density(f5.9$turn.angle[f5.9$Guano=="Absent"], na.rm=T)
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

##Flow = 8.9

##kd1 <- density(f8.9$turn.angle[f8.9$Guano=="Present"], na.rm=T)
##plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='red', lwd=2, main = "Turn Angles Vs Guano, Flow = 8.9")

kd2 <- density(f8.9$turn.angle[f8.9$Guano=="Absent"], na.rm=T)
plot(kd2, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angles Vs Guano, Flow = 8.9")

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)


################################################################

################################ Light ########################

#################################################################


par(mar= c(4,4,2,1))
par(mfrow=c(3,2))
levels(CC.TotalData$Light)

kd1 <- density(f0$turn.angle[f0$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.025), col='yellow', lwd=2, main = "Turn Angles Vs Light, Flow = 0")

kd2 <- density(f0$turn.angle[f0$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)



kd1 <- density(f.6$turn.angle[f.6$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.025), col='yellow', lwd=2, main = "Turn Angles Vs Light, Flow = 0.6")

kd2 <- density(f.6$turn.angle[f.6$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)



kd1 <- density(f3$turn.angle[f3$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.025), col='yellow', lwd=2, main = "Turn Angles Vs Light, Flow = 3")

kd2 <- density(f3$turn.angle[f3$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)



kd1 <- density(f5.9$turn.angle[f5.9$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.025), col='yellow', lwd=2, main = "Turn Angles Vs Light, Flow = 5.9")

kd2 <- density(f5.9$turn.angle[f5.9$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)



kd1 <- density(f8.9$turn.angle[f8.9$Light=="Present"], na.rm=T)
plot(kd1, xlim = c(5,175), ylim = c(0, 0.025), col='yellow', lwd=2, main = "Turn Angles Vs Light, Flow = 8.9")

kd2 <- density(f8.9$turn.angle[f8.9$Light=="Absent"], na.rm=T)
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

```

Turn Angles for each level of Chlorophyll

```{r}

### Kernel density plots ################### Turn Angles
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

################################ Flow Rate ########################
levels(CC.TotalData$Flow.rate)

### CHL 0

kd1 <- density(CHL0$turn.angle[CHL0$Flow.rate=="0"])
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angle Vs Flow Rate, Chlorophyll = 0")

kd2 <- density(CHL0$turn.angle[CHL0$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHL0$turn.angle[CHL0$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHL0$turn.angle[CHL0$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHL0$turn.angle[CHL0$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


### CHL Low

kd1 <- density(CHLL$turn.angle[CHLL$Flow.rate=="0"])
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angle Vs Flow Rate, Chlorophyll = Low")

kd2 <- density(CHLL$turn.angle[CHLL$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLL$turn.angle[CHLL$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHLL$turn.angle[CHLL$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHLL$turn.angle[CHLL$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)



### CHL Medium

kd1 <- density(CHLM$turn.angle[CHLM$Flow.rate=="0"])
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angle Vs Flow Rate, Chlorophyll = Medium")

kd2 <- density(CHLM$turn.angle[CHLM$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLM$turn.angle[CHLM$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHLM$turn.angle[CHLM$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHLM$turn.angle[CHLM$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)



### CHL High

kd1 <- density(CHLH$turn.angle[CHLH$Flow.rate=="0"])
plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angle Vs Flow Rate, Chlorophyll = High")

kd2 <- density(CHLH$turn.angle[CHLH$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLH$turn.angle[CHLH$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

##kd4 <- density(CHLH$turn.angle[CHLH$Flow.rate=="5.9"])
##lines(kd4, col='orange', lwd=2)

##kd5 <- density(CHLH$turn.angle[CHLH$Flow.rate=="8.9"])
##lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)



### CHL Filament

##kd1 <- density(CHLF$turn.angle[CHLF$Flow.rate=="0"])
##plot(kd1, xlim = c(5,175), ylim = c(0, 0.03), col='blue', lwd=2, main = "Turn Angle Vs Flow Rate, Chlorophyll = Filament")

kd2 <- density(CHLF$turn.angle[CHLF$Flow.rate=="0.6"])
plot(kd2, xlim = c(5,175), ylim = c(0, 0.03), col='light blue', lwd=2, main = "Turn Angle Vs Flow Rate, Chlorophyll = Filament")

##kd3 <- density(CHLF$turn.angle[CHLF$Flow.rate=="3"])
##lines(kd3, col='yellow', lwd=2)

##kd4 <- density(CHLF$turn.angle[CHLF$Flow.rate=="5.9"])
##lines(kd4, col='orange', lwd=2)

##kd5 <- density(CHLF$turn.angle[CHLF$Flow.rate=="8.9"])
##lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


###########################################################################

#############################  Light  ######################################

###########################################################################

par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

kd1 <- density(CHL0$turn.angle[CHL0$Light=="Present"])
plot(kd1,xlim = c(5,175), ylim = c(0, 0.02), col='yellow', lwd=2, main = "Turn Angles Vs Light, Chlorophyll = 0")

kd2 <- density(CHL0$turn.angle[CHL0$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


### CHL Low

kd1 <- density(CHLL$turn.angle[CHLL$Light=="Present"])
plot(kd1,xlim = c(5,175), ylim = c(0, 0.02), col='yellow', lwd=2, main = "Turn Angles Vs Light, Chlorophyll = Low")

kd2 <- density(CHLL$turn.angle[CHLL$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


### CHL Medium


kd1 <- density(CHLM$turn.angle[CHLM$Light=="Present"])
plot(kd1,xlim = c(5,175), ylim = c(0, 0.02), col='yellow', lwd=2, main = "Turn Angles Vs Light, Chlorophyll = Medium")

kd2 <- density(CHLM$turn.angle[CHLM$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


### CHL HIgh


kd1 <- density(CHLH$turn.angle[CHLH$Light=="Present"])
plot(kd1,xlim = c(5,175), ylim = c(0, 0.02), col='yellow', lwd=2, main = "Turn Angles Vs Light, Chlorophyll = High")

kd2 <- density(CHLH$turn.angle[CHLH$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


### CHL Filament

kd1 <- density(CHLF$turn.angle[CHLF$Light=="Present"])
plot(kd1,xlim = c(5,175), ylim = c(0, 0.02), col='yellow', lwd=2, main = "Turn Angles Vs Light, Chlorophyll = Filament")

##kd2 <- density(CHLF$turn.angle[CHLF$Light=="Absent"])
##lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

```



Vel.Flow for each level of Flow

```{r}
### Kernel density plots ################### Velocity
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

################################ Chlorophyll ########################

## Flow = 0

kd1 <- density(f0$vel.flow[f0$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-15,15), ylim = c(0, 250), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Chlorophyll, Flow = 0")

kd2 <- density(f0$vel.flow[f0$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f0$vel.flow[f0$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

kd4 <- density(f0$vel.flow[f0$Chlorophyll=="High Chlorophyll"])
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$vel.flow[f0$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)



## Flow = 0.6

kd1 <- density(f.6$vel.flow[f.6$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-15,15), ylim = c(0, 200), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Chlorophyll, Flow = 0.6")

kd2 <- density(f.6$vel.flow[f.6$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f.6$vel.flow[f.6$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

kd4 <- density(f.6$vel.flow[f.6$Chlorophyll=="High Chlorophyll"])
lines(kd4, col='darkgreen', lwd=2)

kd5 <- density(f.6$vel.flow[f.6$Chlorophyll=="Filament"])
lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)




## Flow = 3

kd1 <- density(f3$vel.flow[f3$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 200), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Chlorophyll, Flow = 3")

kd2 <- density(f3$vel.flow[f3$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f3$vel.flow[f3$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

kd4 <- density(f3$vel.flow[f3$Chlorophyll=="High Chlorophyll"])
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f3$vel.flow[f3$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)




## Flow = 5.9

kd1 <- density(f5.9$vel.flow[f5.9$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 200), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Chlorophyll, Flow = 5.9")

kd2 <- density(f5.9$vel.flow[f5.9$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f5.9$vel.flow[f5.9$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

##kd4 <- density(f5.9$vel.flow[f5.9$Chlorophyll=="High Chlorophyll"])
##lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f5.9$vel.flow[f5.9$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)




## Flow = 8.9

kd1 <- density(f8.9$vel.flow[f8.9$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 200), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Chlorophyll, Flow = 8.9")

kd2 <- density(f8.9$vel.flow[f8.9$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f8.9$vel.flow[f8.9$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

##kd4 <- density(f8.9$vel.flow[f8.9$Chlorophyll=="High Chlorophyll"])
##lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f8.9$vel.flow[f8.9$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)



##########################################################################

################################# Guano ##################################

###########################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))
levels(CC.TotalData$Guano)

kd1 <- density(f0$vel.flow[f0$Guano=="Present"])
plot(kd1, xlim = c(-5,5), ylim = c(0, 80), col='red', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Guano, Flow = 0")

kd2 <- density(f0$vel.flow[f0$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

kd1 <- density(f.6$vel.flow[f.6$Guano=="Present"])
plot(kd1, xlim = c(-5,5), ylim = c(0, 80), col='red', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Guano, Flow = 0.6")

kd2 <- density(f.6$vel.flow[f.6$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

kd1 <- density(f3$vel.flow[f3$Guano=="Present"])
plot(kd1, xlim = c(-2,8), ylim = c(0, 80), col='red', lwd=2, main = "Velocity (logged, in relation to flow) Vs Guano, Flow = 3")

kd2 <- density(f3$vel.flow[f3$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

kd1 <- density(f5.9$vel.flow[f5.9$Guano=="Present"])
plot(kd1, xlim = c(1,11), ylim = c(0, 80), col='red', lwd=2, main = "Velocity (logged, in relation to flow) Vs Guano, Flow = 5.9")

kd2 <- density(f5.9$vel.flow[f5.9$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

##kd1 <- density(f8.9$vel.flow[f8.9$Guano=="Present"])
##plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='red', lwd=2, main = "Velocity (logged, in relation to flow) Vs Guano, Flow = 8.9")

kd2 <- density(f8.9$vel.flow[f8.9$Guano=="Absent"])
plot(kd2, xlim = c(4,14), ylim = c(0, 80), col='blue', lwd=2, main = "Velocity (logged, in relation to flow) Vs Guano, Flow = 8.9")
##lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)


##########################################################################

################################# Light ##################################

###########################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))
levels(CC.TotalData$Light)

kd1 <- density(f0$vel.flow[f0$Light=="Present"])
plot(kd1, xlim = c(-5,5), ylim = c(0, 150), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 0")

kd2 <- density(f0$vel.flow[f0$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f.6$vel.flow[f.6$Light=="Present"])
plot(kd1, xlim = c(-5,5), ylim = c(0, 100), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 0.6")

kd2 <- density(f.6$vel.flow[f.6$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='red', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 0.6")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f3$vel.flow[f3$Light=="Present"])
plot(kd1, xlim = c(-2,8), ylim = c(0, 100), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 3")

kd2 <- density(f3$vel.flow[f3$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='red', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 3")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f5.9$vel.flow[f5.9$Light=="Present"])
plot(kd1, xlim = c(1,11), ylim = c(0, 100), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Guano, Flow = 5.9")

kd2 <- density(f5.9$vel.flow[f5.9$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='black', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 5.9")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f8.9$vel.flow[f8.9$Light=="Present"])
plot(kd1, xlim = c(4,14), ylim = c(0, 100), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 8.9")

kd2 <- density(f8.9$vel.flow[f8.9$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='red', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Flow = 8.9")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)



```

Vel.Flow for each level of Chlorophyll

```{r}

### Kernel density plots ################### velocity
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

################################ Flow Rate ########################

## CHL 0
kd1 <- density(CHL0$vel.flow[CHL0$Flow.rate=="0"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Flow Rate, Chlorophyll = 0")

kd2 <- density(CHL0$vel.flow[CHL0$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHL0$vel.flow[CHL0$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHL0$vel.flow[CHL0$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHL0$vel.flow[CHL0$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


## CHL Low

kd1 <- density(CHLL$vel.flow[CHLL$Flow.rate=="0"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Flow Rate, Chlorophyll = Low")

kd2 <- density(CHLL$vel.flow[CHLL$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLL$vel.flow[CHLL$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHLL$vel.flow[CHLL$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHLL$vel.flow[CHLL$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


## CHL Medium

kd1 <- density(CHLM$vel.flow[CHLM$Flow.rate=="0"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Flow Rate, Chlorophyll = Medium")

kd2 <- density(CHLM$vel.flow[CHLM$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLM$vel.flow[CHLM$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHLM$vel.flow[CHLM$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHLM$vel.flow[CHLM$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


## CHL High

kd1 <- density(CHLH$vel.flow[CHLH$Flow.rate=="0"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Flow Rate, Chlorophyll = High")

kd2 <- density(CHLH$vel.flow[CHLH$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLH$vel.flow[CHLH$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

##kd4 <- density(CHLH$vel.flow[CHLH$Flow.rate=="5.9"])
##lines(kd4, col='orange', lwd=2)

##kd5 <- density(CHLH$vel.flow[CHLH$Flow.rate=="8.9"])
##lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)



## CHL Filament

##kd1 <- density(CHLF$vel.flow[CHLF$Flow.rate=="0"])
##plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Flow Rate, Chlorophyll = Filament")

kd2 <- density(CHLF$vel.flow[CHLF$Flow.rate=="0.6"])
plot(kd2, xlim = c(-5,5), ylim = c(0, 50), col='light blue', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Flow Rate, Chlorophyll = Filament")
##lines(kd2, col='light blue', lwd=2)

##kd3 <- density(CHLF$vel.flow[CHLF$Flow.rate=="3"])
##lines(kd3, col='yellow', lwd=2)

##kd4 <- density(CHLF$vel.flow[CHLF$Flow.rate=="5.9"])
##lines(kd4, col='orange', lwd=2)

##kd5 <- density(CHLF$vel.flow[CHLF$Flow.rate=="8.9"])
##lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)

##########################################################################

#############################  Light  ######################################

##########################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

#### CHL 0

kd1 <- density(CHL0$vel.flow[CHL0$Light=="Present"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 30), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Chlorophyll = 0")

kd2 <- density(CHL0$vel.flow[CHL0$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


## CHL Low

kd1 <- density(CHLL$vel.flow[CHLL$Light=="Present"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 120), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Chlorophyll = Low")

kd2 <- density(CHLL$vel.flow[CHLL$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

### CHL Medium

kd1 <- density(CHLM$vel.flow[CHLM$Light=="Present"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 120), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Chlorophyll = Medium")

kd2 <- density(CHLM$vel.flow[CHLM$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


### CHL High

kd1 <- density(CHLH$vel.flow[CHLH$Light=="Present"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 120), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Chlorophyll = High")

kd2 <- density(CHLH$vel.flow[CHLH$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

### CHL Filament

kd1 <- density(CHLF$vel.flow[CHLF$Light=="Present"])
plot(kd1, xlim = c(-15,20), ylim = c(0, 30), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Light, Chlorophyll = Filament")

##kd2 <- density(CHLF$vel.flow[CHLF$Light=="Absent"])
##lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

range(CC.TotalData$turn.angle)
```

Horizontal Headings for each level of Flow

```{r}
### Kernel density plots ################### Velocity
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

################################ Chlorophyll ########################

## Flow = 0

kd1 <- density(f0$heading.pi[f0$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Chlorophyll, Flow = 0")

kd2 <- density(f0$heading.pi[f0$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f0$heading.pi[f0$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

kd4 <- density(f0$heading.pi[f0$Chlorophyll=="High Chlorophyll"])
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f0$heading.pi[f0$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)



## Flow = 0.6

kd1 <- density(f.6$heading.pi[f.6$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Chlorophyll, Flow = 0.6")

kd2 <- density(f.6$heading.pi[f.6$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f.6$heading.pi[f.6$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

kd4 <- density(f.6$heading.pi[f.6$Chlorophyll=="High Chlorophyll"])
lines(kd4, col='darkgreen', lwd=2)

kd5 <- density(f.6$heading.pi[f.6$Chlorophyll=="Filament"])
lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)




## Flow = 3

kd1 <- density(f3$heading.pi[f3$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Chlorophyll, Flow = 3")

kd2 <- density(f3$heading.pi[f3$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f3$heading.pi[f3$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

kd4 <- density(f3$heading.pi[f3$Chlorophyll=="High Chlorophyll"])
lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f3$heading.pi[f3$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)




## Flow = 5.9

kd1 <- density(f5.9$heading.pi[f5.9$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Chlorophyll, Flow = 5.9")

kd2 <- density(f5.9$heading.pi[f5.9$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f5.9$heading.pi[f5.9$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

##kd4 <- density(f5.9$heading.pi[f5.9$Chlorophyll=="High Chlorophyll"])
##lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f5.9$heading.pi[f5.9$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)




## Flow = 8.9

kd1 <- density(f8.9$heading.pi[f8.9$Chlorophyll=="No Chlorophyll"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Chlorophyll, Flow = 8.9")

kd2 <- density(f8.9$heading.pi[f8.9$Chlorophyll=="Low Chlorophyll"])
lines(kd2, col='lightgreen', lwd=2)

kd3 <- density(f8.9$heading.pi[f8.9$Chlorophyll=="Medium Chlorophyll"])
lines(kd3, col='green', lwd=2)

##kd4 <- density(f8.9$heading.pi[f8.9$Chlorophyll=="High Chlorophyll"])
##lines(kd4, col='darkgreen', lwd=2)

##kd5 <- density(f8.9$heading.pi[f8.9$Chlorophyll=="Filament"])
##lines(kd5, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("No Chlorophyll", "Low Chlorophyll", "Medium Chlorophyll", "High Chlorophyll", "Filament"),
       col = c("blue", "lightgreen", "green", "darkgreen", "black"),
       lty = 1)



##########################################################################

################################# Guano ##################################

###########################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))
levels(CC.TotalData$Guano)

kd1 <- density(f0$heading.pi[f0$Guano=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='red', lwd=2, main = "Heading Vs Guano, Flow = 0")

kd2 <- density(f0$heading.pi[f0$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

kd1 <- density(f.6$heading.pi[f.6$Guano=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='red', lwd=2, main = "Heading Vs Guano, Flow = 0.6")

kd2 <- density(f.6$heading.pi[f.6$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

kd1 <- density(f3$heading.pi[f3$Guano=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='red', lwd=2, main = "Heading Vs Guano, Flow = 3")

kd2 <- density(f3$heading.pi[f3$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

kd1 <- density(f5.9$heading.pi[f5.9$Guano=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='red', lwd=2, main = "Heading Vs Guano, Flow = 5.9")

kd2 <- density(f5.9$heading.pi[f5.9$Guano=="Absent"])
lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)

##kd1 <- density(f8.9$heading.pi[f8.9$Guano=="Present"])
##plot(kd1, xlim = c(-15,20), ylim = c(0, 250), col='red', lwd=2, main = "Heading Vs Guano, Flow = 8.9")

kd2 <- density(f8.9$heading.pi[f8.9$Guano=="Absent"])
plot(kd2, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Guano, Flow = 8.9")
##lines(kd2, col='blue', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Guano Present", "Guano Absent"),
       col = c("red", "blue"),
       lty = 1)


##########################################################################

################################# Light ##################################

###########################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))
levels(CC.TotalData$Light)

kd1 <- density(f0$heading.pi[f0$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.05), col='yellow', lwd=2, main = "Heading Vs Light, Flow = 0")

kd2 <- density(f0$heading.pi[f0$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f.6$heading.pi[f.6$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.05), col='yellow', lwd=2, main = "Heading Vs Light, Flow = 0.6")

kd2 <- density(f.6$heading.pi[f.6$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='red', lwd=2, main = "Heading Vs Light, Flow = 0.6")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f3$heading.pi[f3$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.05), col='yellow', lwd=2, main = "Heading Vs Light, Flow = 3")

kd2 <- density(f3$heading.pi[f3$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='red', lwd=2, main = "Heading Vs Light, Flow = 3")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f5.9$heading.pi[f5.9$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.05), col='yellow', lwd=2, main = "Velocity (not logged, in relation to flow) Vs Guano, Flow = 5.9")

kd2 <- density(f5.9$heading.pi[f5.9$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='black', lwd=2, main = "Heading Vs Light, Flow = 5.9")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

kd1 <- density(f8.9$heading.pi[f8.9$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.05), col='yellow', lwd=2, main = "Heading Vs Light, Flow = 8.9")

kd2 <- density(f8.9$heading.pi[f8.9$Light=="Absent"])
##plot(kd2, xlim = c(-10,10), ylim = c(0, 25), col='red', lwd=2, main = "Heading Vs Light, Flow = 8.9")
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


```

Horizontal Headings for each level of Chlorophyll

```{r}

### Kernel density plots ################### Heading
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

################################ Flow Rate ########################

## CHL 0
kd1 <- density(CHL0$heading.pi[CHL0$Flow.rate=="0"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Flow Rate, Chlorophyll = 0")

kd2 <- density(CHL0$heading.pi[CHL0$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHL0$heading.pi[CHL0$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHL0$heading.pi[CHL0$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHL0$heading.pi[CHL0$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


## CHL Low

kd1 <- density(CHLL$heading.pi[CHLL$Flow.rate=="0"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Flow Rate, Chlorophyll = Low")

kd2 <- density(CHLL$heading.pi[CHLL$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLL$heading.pi[CHLL$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHLL$heading.pi[CHLL$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHLL$heading.pi[CHLL$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


## CHL Medium

kd1 <- density(CHLM$heading.pi[CHLM$Flow.rate=="0"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Flow Rate, Chlorophyll = Medium")

kd2 <- density(CHLM$heading.pi[CHLM$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLM$heading.pi[CHLM$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

kd4 <- density(CHLM$heading.pi[CHLM$Flow.rate=="5.9"])
lines(kd4, col='orange', lwd=2)

kd5 <- density(CHLM$heading.pi[CHLM$Flow.rate=="8.9"])
lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)


## CHL High

kd1 <- density(CHLH$heading.pi[CHLH$Flow.rate=="0"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Flow Rate, Chlorophyll = High")

kd2 <- density(CHLH$heading.pi[CHLH$Flow.rate=="0.6"])
lines(kd2, col='light blue', lwd=2)

kd3 <- density(CHLH$heading.pi[CHLH$Flow.rate=="3"])
lines(kd3, col='yellow', lwd=2)

##kd4 <- density(CHLH$heading.pi[CHLH$Flow.rate=="5.9"])
##lines(kd4, col='orange', lwd=2)

##kd5 <- density(CHLH$heading.pi[CHLH$Flow.rate=="8.9"])
##lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)



## CHL Filament

##kd1 <- density(CHLF$heading.pi[CHLF$Flow.rate=="0"])
##plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='blue', lwd=2, main = "Heading Vs Flow Rate, Chlorophyll = Filament")

kd2 <- density(CHLF$heading.pi[CHLF$Flow.rate=="0.6"])
plot(kd2, xlim = c(-190,190), ylim = c(0, 0.03), col='light blue', lwd=2, main = "Heading Vs Flow Rate, Chlorophyll = Filament")
##lines(kd2, col='light blue', lwd=2)

##kd3 <- density(CHLF$heading.pi[CHLF$Flow.rate=="3"])
##lines(kd3, col='yellow', lwd=2)

##kd4 <- density(CHLF$heading.pi[CHLF$Flow.rate=="5.9"])
##lines(kd4, col='orange', lwd=2)

##kd5 <- density(CHLF$heading.pi[CHLF$Flow.rate=="8.9"])
##lines(kd5, col='red', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Flow rate  0", "Flow rate 0.6", "Flow rate 3", "Flow rate 5.9", "Flow rate 8.9"),
       col = c("blue", "light blue", "yellow", "orange", "red"),
       lty = 1)

##########################################################################

#############################  Light  ######################################

##########################################################################
par(mar= c(4,4,2,1))
par(mfrow=c(3,2))

#### CHL 0

kd1 <- density(CHL0$heading.pi[CHL0$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='yellow', lwd=2, main = "Heading Vs Light, Chlorophyll = 0")

kd2 <- density(CHL0$heading.pi[CHL0$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


## CHL Low

kd1 <- density(CHLL$heading.pi[CHLL$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='yellow', lwd=2, main = "Heading Vs Light, Chlorophyll = Low")

kd2 <- density(CHLL$heading.pi[CHLL$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

### CHL Medium

kd1 <- density(CHLM$heading.pi[CHLM$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='yellow', lwd=2, main = "Heading Vs Light, Chlorophyll = Medium")

kd2 <- density(CHLM$heading.pi[CHLM$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


### CHL High

kd1 <- density(CHLH$heading.pi[CHLH$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='yellow', lwd=2, main = "Heading Vs Light, Chlorophyll = High")

kd2 <- density(CHLH$heading.pi[CHLH$Light=="Absent"])
lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)

### CHL Filament

kd1 <- density(CHLF$heading.pi[CHLF$Light=="Present"])
plot(kd1, xlim = c(-190,190), ylim = c(0, 0.03), col='yellow', lwd=2, main = "Heading Vs Light, Chlorophyll = Filament")

##kd2 <- density(CHLF$heading.pi[CHLF$Light=="Absent"])
##lines(kd2, col='black', lwd=2)

legend("topright",                                  # Add legend to density
       legend = c("Light Present", "Light Absent"),
       col = c("yellow", "black"),
       lty = 1)


```

Using subset data from notebook 11
Plotting histograms for each condition level, starting with turn angles, Horizontal and Vertical Heading, and Velocity next...
```{r}
par(mar= c(4,4,2,1))
par(mfrow=c(1,3))


hist_info <- hist(DHN$turn.anglexysmooth, plot = FALSE)         # Store output of hist function
hist_info$density <- hist_info$counts /    # Compute density values
  sum(hist_info$counts) * 100
plot(hist_info, freq = FALSE, xlab = "Turn Angle (degrees)", ylab = "Percentage (%)", main = "Dark, High Flow (5.9 cm s-1), No Chemicals", xlim=c(0,100), ylim=c(0,75), xaxp = c(0, 180, 18))              # Plot histogram with percentages

hist_info <- hist(DHG$turn.anglexysmooth, plot = FALSE)         # Store output of hist function
hist_info$density <- hist_info$counts /    # Compute density values
  sum(hist_info$counts) * 100
plot(hist_info, freq = FALSE, xlab = "Turn Angle (degrees)", ylab = "Percentage (%)", main = "Dark, High Flow (5.9 cm s-1), Guano Only", xlim=c(0,100), ylim=c(0,75), xaxp = c(0, 180, 18))              # Plot histogram with percentages

hist_info <- hist(DHC$turn.anglexysmooth, plot = FALSE)         # Store output of hist function
hist_info$density <- hist_info$counts /    # Compute density values
  sum(hist_info$counts) * 100
plot(hist_info, freq = FALSE, xlab = "Turn Angle (degrees)", ylab = "Percentage (%)", main = "Dark, High Flow (5.9 cm s-1), Guano and Chlorophyll", xlim=c(0,100), ylim=c(0,75), xaxp = c(0, 180, 18))              # Plot histogram with percentages

save.image("~/Post-doc/Data/Total Merged Data File (Sep 14 2023).RData")
```



Circular stats for pathwise means of angular data


```{r}
library(tidyverse)
library(circular)
library(bbmle)

dvm <- circular:::DvonmisesRad
## check vectorization
stopifnot(all.equal(c(dvm(1,1,0.5,log=TRUE), dvm(2,3,0.5,log=TRUE)),
                    dvm(c(1,2), c(1,3), 0.5, log=TRUE)))

range(AGG_Dark$pitch.perfect)

m_null <- mle2(pitch.perfect ~ dvm(exp(log_mu), exp(log_kappa)),
           parameters = list(log_mu ~ 1),
           data = AGG_Dark,
           start = list(log_mu = 0, log_kappa = 0))

m3_FC <- update(m_null, parameters = list(log_mu ~ Flow.rate+Chem))
m3_F  <- update(m_null, parameters = list(log_mu ~ Flow.rate))


anova(m3_FC, m3_F, m_null)
```
