==> DataProcessing/Notebooks/notebook12-autocorrelation-distributions-for-model.Rmd <==
---
title: "notebook12-autocorrelation-distribution-for-models"
output:
  html_document:
    df_print: paged
---

Load data
```{r echo=F}
rm(list=ls(all=TRUE))
#load("~/Post-doc/Data/Total Merged Data File (Sep 2 2022).RData")
#load('~/Work/Data/Antarctic/CC.TotalData.2022.07.20.RData')
CC.TotalData <- read.csv('~/Bigelow/Data/CC.TotalData.July.15.2024.csv', header = T)
library(ggplot2)
library(GGally)
library(hexbin)
library(diptest)
library(tidymodels)
library(randomForest)

str(CC.TotalData)
CC.TotalData$Chlorophyll <- as.factor(CC.TotalData$Chlorophyll)
levels(CC.TotalData$Chlorophyll)
```

```{r}
# Values of conditions to test
flow.rate <- 0  # 0 0.6 3 5.9 8.9
chlorophyll <- 0 # 0 4.3 4.6 5.5 6.1 7.6 13.5 19
guano <- 1 # Absent=1 Present=2
light <- 1 # Absent=1 Present=2

# Vectors of all possible conditions combinations
frs <- as.numeric(as.character(unique(CC.TotalData$Flow.rate)))
chls <- as.numeric(as.character(unique(CC.TotalData$Chlorophyll)))
chls <- na.omit(chls)
guans <- c(1,2)
CC.TotalData$Guano <- as.factor(CC.TotalData$Guano)
levels(CC.TotalData$Guano)
levels(CC.TotalData$Guano) <- c("1","2")
lights <- c(1,2)
CC.TotalData$Light <- as.factor(CC.TotalData$Light)
levels(CC.TotalData$Light)
levels(CC.TotalData$Light) <- c("1","2")
conditions <- expand.grid(frs,chls,guans,lights)
```

Autocorrelation looking at linear regression and residuals. Note that residuals are not normally distributed.
```{r}
velocity <- CC.TotalData$v[
  (CC.TotalData$Flow.rate== "0" & 
     CC.TotalData$Chlorophyll== "0" & 
     CC.TotalData$Guano== "Absent" &
     CC.TotalData$Light== "Present")]

x <- log10(velocity[1:length(velocity)-1])
y <- log10(velocity[2:length(velocity)])


vel.auto.lm <- lm(x ~ y + 1)
summary(vel.auto.lm)

plot(hexbin(x,
            y,
            xbins = 100),
     xlab='Log velocity (m/s)',
     ylab='Log velocity (m/s) next time step',
     legend = 0)

hist(log10(velocity),100)
hist(vel.auto.lm$residuals,100)
qqnorm(vel.auto.lm$residuals,main="Normal Q-Q plot")
qqline(vel.auto.lm$residuals)
ks.test(vel.auto.lm$residuals,'pnorm')

```


Loop through combinations

```{r}
for (i in 1:dim(conditions)[1])
{
  velocity <- CC.TotalData$v[
  (CC.TotalData$Flow.rate==conditions[i,1] & 
     CC.TotalData$Chlorophyll==conditions[i,2] & 
     as.numeric(CC.TotalData$Guano)==conditions[i,3] &
     as.numeric(CC.TotalData$Light)==conditions[i,4])]
  if (length(velocity) <= 1) {
    conditions[i,5] <- NA } else {
      c<-cor.test(log10(velocity[1:length(velocity)-1]),
                  log10(velocity[2:length(velocity)]))
      conditions[i,5] <- c$estimate
        v0 <- log10(velocity[1:length(velocity)-1])
        v1 <- log10(velocity[2:length(velocity)])
        df <- data.frame(v0,v1)
        fit <- lm(v1 ~ v0 + 1, data = df)
        conditions[i,6] <- fit$coefficients[1]
        conditions[i,7] <- fit$coefficients[2]
        d <- dip.test(velocity)
        conditions[i,8] <- d$p.value
        hist(log10(velocity),100,
             main=d$p.value)
  }  
}
conditions <- setNames(conditions,c("flow.rate","chlorophyll","guano","light","corr.val","slope","intercept","dip.test"))
```


Using random forest to fit the missing values

```{r}
Idata <- which(!is.na(conditions[,5])) # Index of data
Iblank <- which(is.na(conditions[,5])) # Index of blank data

conditions_data <- conditions[Idata,]
conditions_blank <- conditions[Iblank,]


# Tidy Models version
#data_split <- initial_split(conditions_data, prop = 0.6)
#data_test <- testing(data_split)
#data_train <- training(data_split)
#data_folds <- vfold_cv(data_train, 
#                       v = 5, # number of partitions in dataset 
#                       repeats = 1)
#mc_split <- mc_cv(data_train, 
#                  prop = 9/10, # proportion to use per resample
#                  times = 20) # number of resamples
#template_data <- training(data_split)


# Random forest package version - doing dip test
conditions.rf <- randomForest(dip.test ~ .,
                              data=select(conditions_data,c(1,2,3,4,8)),
                              importance=TRUE,
                              proximity=TRUE)
#print(conditions.rf)
#round(importance(conditions.rf), 2)
#sqrt(conditions.rf$mse[which.min(conditions.rf$mse)]) 
plot(conditions.rf)
varImpPlot(conditions.rf)
conditions_blank$dip.test <- predict(conditions.rf,conditions_blank[,1:4])
fit.dip.test <- predict(conditions.rf,conditions_data[,1:4])
plot(fit.dip.test,conditions_data$dip.test)
plot(conditions_blank$chlorophyll,conditions_blank$dip.test)
plot(conditions_blank$flow.rate,conditions_blank$dip.test)
plot(conditions_blank$light,conditions_blank$dip.test)
plot(conditions_blank$guano,conditions_blank$dip.test)

```

Random forest test for corr
```{r}
conditions.rf <- randomForest(corr.val ~ .,
                              data=select(conditions_data,c(1,2,3,4,5)),
                              importance=TRUE,
                              proximity=TRUE)
#print(conditions.rf)
#round(importance(conditions.rf), 2)
#sqrt(conditions.rf$mse[which.min(conditions.rf$mse)]) 
plot(conditions.rf)
varImpPlot(conditions.rf)
conditions_blank$corr.val <- predict(conditions.rf,conditions_blank[,1:4])
fit.corr.val <- predict(conditions.rf,conditions_data[,1:4])
plot(fit.corr.val,conditions_data$corr.val)
plot(conditions_blank$chlorophyll,conditions_blank$corr.val)
plot(conditions_blank$flow.rate,conditions_blank$corr.val)
plot(conditions_blank$light,conditions_blank$corr.val)
plot(conditions_blank$guano,conditions_blank$corr.val)
```
