==> DataProcessing/Notebooks/notebook14-simulation.Rmd <==
---
title: "notebook14-simulation"
output:
  html_document:
    df_print: paged
---

```{r echo=F}
#knitr::opts_chunk$set(warning = FALSE, message = FALSE)
source('../functions/model-source.R')
library('plotly')
```

A quick example of a series of simulations in a tank
```{r echo=F, warning = F, message=F}
paths <- as.numeric(c(NA,NA,NA))
for (i in 1:20)
{
  dataout <- krilltankinit(nt=1000,xsize=25000,ysize=33000,zsize=18000)
  paths <- rbind(paths,dataout)
}
krilltankplot(paths)
```

A different example
```{r echo=F, warning = F, message=F}
nk <- 500
swarm <- matrix(data=NA,nrow=nk,ncol=3)
for (i in 1:nk)
{
  dataout <- krilltankinit2(nt=1000,
                            xsize=25000,ysize=33000,zsize=18000,
                            xi=12000,yi=20000,zi=9000,
                            flow.rate=100, chloro = 50, guano = 1, light=1)
  swarm[i,] <- dataout[1000,]
}
fig <- plot_ly(x=swarm[,1], y=swarm[,2], z=swarm[,3])
fig
```
