==> DataProcessing/Notebooks/notebook14b-notes for Nicole.Rmd <==
Setwd (‘~Nicole/Documents/Post-doc/krill-tank-code/Notebooks/’)
krilltankinit <- function(xsize = 256,  ## tank size
                          ysize = 332,  ## tank size
                          zsize = 179,  ## tank size
                          xi = 100, ## krill starting position
                          yi = 100, ## krill starting position
                          zi = 100, ## krill starting position
                          psi = 0,  ## starting orientation horizontal
                          theta = 0,  ## starting orientation vertical
                          v = 0.1,  ## krill starting velocity
                          nt = 10000, ## number of time steps
                          dt = 1) ## time interval
{
  dataout <- matrix(data=NA,nrow=nt,ncol=3)  ## getting the XYZ columns for the krill?
  dataout[1,1:3] <- c(xi,yi,zi) ## creating the starting location as the first row of XYZ?
  
  mu <- 0 ## mean of distribution
  sigma <- 0.5 ## standard deviation in distribution
  
for (i in 1:(nt-1))
  {
    v <- v + rnorm(1, mean = mu, sd = sigma)/100
    #v=v+(mu+sigma*tan(pi*(rand(1,1)-1/2)))/100; # Cauchy (matlab)
    psi <- psi + (runif(1)-.5)*5;
    theta <- theta + (runif(1)-.5)*5;
    dataout[i+1,1] <- dataout[i,1] + v * dt * cos(pi*psi/180)
    dataout[i+1,2] <- dataout[i,2] + v * dt * sin(pi*psi/180)
    dataout[i+1,3] <- dataout[i,3] + v * dt * sin(pi*theta/180)
    dataout[i+1,1] <- max(min(dataout[i+1,1],xsize),0)
    dataout[i+1,2] <- max(min(dataout[i+1,2],ysize),0)
    dataout[i+1,3] <- max(min(dataout[i+1,3],zsize),0)
  }
  return(dataout)  ## returns XYZ of paths
}

krilltankinit2 <- function(xsize = 256,
                          ysize = 332,
                          zsize = 179,
                          xi = 100,
                          yi = 100,
                          zi = 100,
                          psi = 0,
                          theta = 0,
                          v = 0.1,
                          nt = 10000,
                          dt = 1,
                          flow.rate = 0, ## adding parameters
                          chloro = 0,
                          guano = 0,
                          light = 0,
                          filein = 'notebook13-rf-24.07.RData')  ##choosing data to add
{
  dataout <- matrix(data=NA,nrow=nt,ncol=3)
  dataout[1,1:3] <- c(xi,yi,zi)
  
  params <- getparams(filein = filein,
                      flow.rate = flow.rate,
                      chloro = chloro,
                      guano = guano,
                      light = light)
  slope <- params[1]  ## see figure below
  intercept <- params[2] ##autocorrelation parameters for velocity from data
  sigma <- params[3]
  #mu <- 0
  #sigma <- 0.5
  
  
  for (i in 1:(nt-1))
  {
    v <- v * slope + intercept + rnorm(1, mean = 0, sd = sigma)
    #v <- v + rnorm(1, mean = mu, sd = sigma)/100
    #v=v+(mu+sigma*tan(pi*(rand(1,1)-1/2)))/100; # Cauchy (matlab)
    psi <- psi + (runif(1)-.5)*5;
    theta <- theta + (runif(1)-.5)*5;
    dataout[i+1,1] <- dataout[i,1] + v * dt * cos(pi*psi/180)
    dataout[i+1,2] <- dataout[i,2] + v * dt * sin(pi*psi/180)
    dataout[i+1,3] <- dataout[i,3] + v * dt * sin(pi*theta/180)
    dataout[i+1,1] <- max(min(dataout[i+1,1],xsize),0)
    dataout[i+1,2] <- max(min(dataout[i+1,2],ysize),0)
    dataout[i+1,3] <- max(min(dataout[i+1,3],zsize),0)
  }
  return(dataout)
}

## plotting XYZ of krill in model
krilltankplot <- function(datain = c(NA,NA,NA))
{
  fig <- plot_ly(x=datain[,1], y=datain[,2], z=datain[,3],
                 mode='lines')
  fig
  return(fig)
}

getparams <- function(filein = 'notebook13-rf-24.07.2023data.RData',
                      flow.rate = 0,
                      chloro = 0,
                      guano = 0,
                      light = 0)
{
  load(filein)  ## find fit of track compared to data distributions
  fit.slope <- predict(conditions.rf.slope,
                       c(flow.rate,chloro,guano,light))
  fit.intercept <- predict(conditions.rf.intercept,
                       c(flow.rate,chloro,guano,light))
  fit.sigma <- predict(conditions.rf.sigma,
                       c(flow.rate,chloro,guano,light))
  return(c(fit.slope,fit.intercept,fit.sigma))
}

notebook14-simulations

source('../functions/model-source.R') ## Pulls the model from model-source.R
library('plotly')

xyz <- krilltankinit() ## creates matrix of krill track XYZs, change nt, xsize, ysize, size here. E.g. nt=10000, xsize=25000, ysize=33000, zsize = 18000 (NOT PULLING FROM DATA)
krilltankplot(xyz)
krilltankplot(krilltankinit())  ##  will create a new track path and plot it each time you run it



xyz2 <- krilltankinit2(nt=10000, xsize=25000, ysize=33000, zsize=18000) ## as above but uses data from experiments
krilltankplot(xyz2)


A quick example of a series of simulations in a tank

paths <- as.numeric(c(NA,NA,NA))  ## creates matrix for multiple paths
for (i in 1:20)  ## generates 20 paths
{
  dataout <- krilltankinit(nt=10000, xsize=25000, ysize=33000, zsize=18000)  ## currently not pulling from data, change to krilltankinit2 to pull from data
  paths <- rbind(paths,dataout)
}
krilltankplot(paths)  ## plots all 20 paths



A different Example
nk <- 100  ## number of krill

swarm <- matrix(data=NA,nrow=nk,ncol=3) ## creates matrix for large number of krill
for (i in 1:nk)
{
  dataout <- krilltankinit2(nt=1000,
                            xsize=25000,ysize=33000,zsize=18000,
                            xi=12000,yi=20000,zi=9000,
                            flow.rate=100)
  swarm[i,] <- dataout[1000,]
}

plot(swarm[,1], swarm[,2])  ##plots final position of all 100 krill in the swarm after all time steps

krilltankplot(...) ## need to update from github

fig <- plot(krilltankplot())

fig
