==> DataProcessing/Notebooks/notebook15-ngdr.Rmd <==
---
title: "NGDR"
output: html_notebook
---
NGDR is the ratio of the net displacement (linear distance between starting point and end point of a path) to the gross displacement (total distance travelled) for each path recorded. 


```{r}
rm(list=ls(all=TRUE))   ## removes the previous workspace and environment so that we only have the data we need loaded in the session
AGG_Dark <- read.csv("~/Bigelow/Data/AGG_Dark_Marc.csv")
head(AGG_Dark)

#FeedingDark <- read.csv(file = "~/Bigelow/Manuscripts/Feeding MS/FeedingDark.csv", sep = ',', header = T)
AGG_Dark<- read.csv(file = "~/Bigelow/Manuscripts/Feeding MS/Data/AGG_Dark.csv", sep = ',', header = T)
FeedingDark<- read.csv(file = "~/Bigelow/Manuscripts/Feeding MS/Data/FeedingDark.csv", sep = ',', header = T)
```

Calculate Total distance of track
```{r}
library(dplyr)
head(Feeding)

total.dist <- function(x,y) {
  sum(sqrt((x-lag(x))^2+(y-lag(y))^2),na.rm=TRUE)
}

res <- Feeding %>% group_by(D_V_T) %>% summarise(dtot=total.dist(xsmooth,ysmooth))

print(res)

NGDR <- merge(AGG_Feed, res)

head(NGDR)

write.csv(NGDR, file = "~/Post-doc/Data/NGDR Sep 29 2023.csv", row.names = FALSE)


```

Calculate distance from start x,y to end, x,y location
```{r}
full_dist <- data.frame()
for (dvt in unique(Feeding$D_V_T)) {
  data1 <- Feeding %>% subset(D_V_T==dvt)
    sd1 <- data1[1,"smooth.d"]
    sd2 <- data1[5000,"smooth.d"]
    diff.dist <- sd1-sd2
    tmp <- data.frame(D_V_T=dvt,i,sd1,sd2,diff.dist)
    full_dist <- rbind(full_dist,tmp)
    print(dvt)
}


str(full_dist)
NGDR2 <- na.omit(NGDR2)
full_dist %>% group_by(D_V_T) %>% summarise_at(vars(diff.dist), mean)

NGDR2 <- merge(NGDR, full_dist)

head(NGDR2)
tail(NGDR2)

write.csv(NGDR2, file = "~/Post-doc/Data/NGDR Sep 29 2023.csv", row.names = FALSE)

```

Calculate NGDR ratio
```{r}
NGDR2$NGDR <- NGDR2$diff.dist/NGDR2$dtot
head(NGDR2)

range(NGDR2$diff.dist)

NGDR2 %>% group_by(Light, Flow.rate, Chem) %>% summarise_at(vars(NGDR, diff.dist, dtot), mean)
```

Plotting Angular data into histograms on polar projection
```{r}
## redo with vertical headings, change range for turn angles?

library(ggplot2)
head(AGG_Feed)
str(AGG_Feed)
range(AGG_Feed$turn.anglexysmooth)
AGG_Feed$turn.anglexysmooth <- AGG_Feed$turn.anglexysmooth/(2*pi)*360
f_mu <- mean(Feeding$turn.angle.smooth, by =c(Chem, Light, Flow.rate))

ggplot(AGG_Dark, aes(heading.smooth.degrees, fill = Chemical.Cue)) +
  geom_histogram(binwidth = 15, boundary = 0, colour = "black", size = .25) +
  coord_polar() +
   ylim(0, 12)  +
  scale_x_continuous(limits = c(0,360),
                     breaks = seq(0, 360, by = 45),
                     minor_breaks = seq(0, 360, by = 15))+
  theme_classic()+
  theme(axis.text=element_text(color="black"))+
  theme(text=element_text(size=20))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  theme(legend.key = element_rect(fill = "white"))+
  ##geom_vline(aes(heading.pi), xintercept = f_mu)+   ## adds mean angle line to plot
  labs(size = "", x = "Heading Angle (degrees)", y = "Frequency of Heading Angle", title = "")+
  scale_fill_manual(values=c("white", "gray", "gray35"),name = "Chemical Cue",
                    labels=c("Ambient", "Guano", "Guano+CHL"))+
 facet_grid(facets = Flow.Rate ~ Chemical.Cue, labeller = label_both)


################# turn angles arrow plot
# plot 
n <- 128

ggplot(data = dat, aes(x = turn.anglexysmooth, colour = Light)) +
  geom_spoke(data = dat, aes(x = xsmooth, y = ysmooth, angle = turn.anglexysmooth, radius = n/2),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  geom_spoke(data = dat, aes(x = xsmooth, y = ysmooth, angle = turn.anglexysmooth+30, radius = n/2),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  geom_spoke(data = dat, aes(x = xsmooth, y = ysmooth, angle = turn.anglexysmooth+35, radius = n/2),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  coord_equal()+
  ylim(0, 90)+
  scale_x_continuous(limits = c(0,90),
                     breaks = seq(0, 90, by = 45),
                     minor_breaks = seq(0, 360, by = 15))+
  ##scale_y_continuous(limits = c(0,90),
                    ## breaks = seq(0, 90, by = 15),
                    ## minor_breaks = seq(0, 90, by = 5))+
  theme_classic()+
  scale_color_manual(values = c("black", "gray"))+
  labs(size = "", x = "Coefficient of Dispersion", y = "Turn Angle (degrees)", title = "") +
  facet_grid(facets = Chem ~ Flow, labeller = label_both)

#############################################

################### Horiz heading plot


dat$heading.pi+360
dat$heading.pi


##works but want it in co-ord polar and with arrows
ggplot(dat, aes(heading.pi, col=Light))+
  geom_point(data = dat, aes(y = heading.pi))+
  facet_grid(facets = Chem ~ Flow, labeller = label_both)



## adding each element in
ggplot(dat, aes(heading.pi, col=Light))+
  geom_point(data = dat, aes(y = heading.pi))+
  scale_color_manual(values = c("black", "gray"))+
  labs(size = "", x = "Coefficient of Dispersion", y = "Horizontal Heading (degrees)", title =     "") +
  facet_grid(facets = Chem ~ Flow, labeller = label_both)

##
ggplot(dat, aes(heading.pi, col=Light))+
  geom_point(data = dat, aes(y = 0))+
  xlim(-90,90)+
  ylim(-.5, .5)+
  theme_classic()+
  scale_color_manual(values = c("black", "gray"))+
  labs(size = "", x = "Horizontal Heading (degrees)", y = "Coefficient of Dispersion", title =     "") +
  facet_grid(facets = Chem ~ Flow, labeller = label_both)

# plot 
n <- 128

##need to center y axis to x axis
ggplot(data = dat, aes(x = heading.pi, y = ysmooth, colour = Light)) +
  geom_spoke(data = dat, aes(x = 0, y = 0, angle = heading.pi, radius = n/2),         arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  geom_spoke(data = dat, aes(x = 0, y = 0, angle = heading.pi+360, radius = n/2),         arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  coord_polar() +
  scale_x_continuous(limits = c(0,360),
                     breaks = seq(0, 360, by = 45),
                     minor_breaks = seq(0, 360, by = 15))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  theme(legend.key = element_rect(fill = "white"))+
  scale_color_manual(values = c("black", "gray"))+
  labs(size = "", x = "Coefficient of Dispersion", y = "Horizontal Heading (degrees)", title =     "") +
  facet_grid(facets = Chem ~ Flow, labeller = label_both)

###############################################################################

## trying in coord_equal instead

ggplot(data = dat, aes(x = heading.pi, colour = Light)) +
  geom_spoke(data = dat, aes(x = ysmooth, y = xsmooth, angle = heading.pi*(-1), radius = n/2),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  geom_spoke(data = dat, aes(x = ysmooth, y = xsmooth, angle = heading.pi+90, radius = n/2),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  geom_spoke(data = dat, aes(x = ysmooth, y = xsmooth, angle = heading.pi+180, radius = n/2),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  coord_equal()+
  ylim(0, 90)+
  scale_x_continuous(limits = c(0,90),
                     breaks = seq(0, 90, by = 45),
                     minor_breaks = seq(0, 90, by = 15))+
  ##scale_y_continuous(limits = c(0,90),
                    ## breaks = seq(0, 90, by = 15),
                    ## minor_breaks = seq(0, 90, by = 5))+
  theme_classic()+
  scale_color_manual(values = c("black", "gray"))+
  labs(size = "", x = "Heading Angle (degrees)", y = "Coefficient of Dispersion", title = "") +
  facet_grid(facets = Chem ~ Flow, labeller = label_both)


```

```{r}
#######################################################################################
#### angular plots with mean heading as arrow and length as coefficient of dispersion
library(ggplot2)

dat <- aggregate(AGG_Feed, by = list(AGG_Feed$Flow.rate, AGG_Feed$Light, AGG_Feed$Chem), FUN=mean)
colnames(dat) <- c("Flow", "Light", "Chem", "DVT", "Guano", "Light2", "Chem2", "Flow2", "X", "Y", "Z", "dx", "dy", "dz", "d", "vx", "vy", "vz", "v", "h.heading", "v.heading", "turn.anglexy", "turn.angleyz", "turn.angle", "vel.turn.angle", "vel","vel.flow", "trim.X", "xsmooth", "ysmooth", "zsmooth", "smooth.dx", "smooth.dy", "smooth.dz", "smooth.d",  "smooth.vx", "smooth.vy", "smooth.vz", "smooth.v",  "smooth.h.heading", "heading.pi", "smooth.v.heading", "pitch.perfect", "turn.anglexysmooth", "turn.angleyzsmooth")

levels(dat$Chem)
levels(dat$Chem) <- c("Ambient", "Guano", "Guano+CHL")

n <- nrow(dat)
range(AGG_Feed$turn.anglexysmooth)


write.csv(dat, file = "~/Post-doc/Data/Aggregated means(dat).csv", row.names = FALSE)

# plot (not working as intended but close)
ggplot(data = AGG_Feed, aes(x = xsmooth, y = ysmooth, colour = Light)) +
  geom_spoke(data = AGG_Feed, aes(x = xsmooth, y = ysmooth, angle = turn.anglexysmooth+30, radius = 12),       arrow=arrow(length = unit(0.2,"cm")), size = 1) +
  coord_equal()+
  ##scale_x_continuous(limits = c(0,90),
                    ## breaks = seq(0, 90, by = 15),
                    ## minor_breaks = seq(0, 90, by = 5))+
  ##scale_y_continuous(limits = c(0,90),
                    ## breaks = seq(0, 90, by = 15),
                    ## minor_breaks = seq(0, 90, by = 5))+
  theme_classic()+
  scale_color_manual(values = c("gray", "black"))+
  labs(size = "", x = "X", y = "Y", title = "") +
  facet_grid(facets = Chem ~ Flow.rate, labeller = label_both)

###########################################################################################
library(MASS)
eqscplot(dat$xsmooth, dat$ysmooth, xlim=c(0,0.5), ylim=c(0,0.5), xlab = "X", ylab = "Y", main = "", axes=FALSE, frame.plot = FALSE)
axis(1, pos=0)
axis(2, pos=0)
arrows(0,0,x1=dat$xsmooth, y1 = dat$ysmooth, col = dat$Light)
symbols(0,0,2, add = TRUE, inches = FALSE, asp = 1)

##############################################################
      ##### Heading plots   ######

library(ggmulti)
library(forcats)
library(ggplot2)
library(dplyr)

range(AGG_Dark$turn.angle)
str(AGG_Dark)
AGG_Dark$D_V_T <- as.factor(AGG_Dark$D_V_T)
AGG_Dark$Chem <- as.factor(AGG_Dark$Chem)
AGG_Dark$Flow.rate <- as.factor(AGG_Dark$Flow.rate)
levels(AGG_Dark$Flow.rate)
AGG_Dark$Flow.rate <- relevel(AGG_Dark$Flow.rate, "Medium Flow")

AGG_Dark %>%
  filter(Chem %in% c("Ambient", "Guano", "Guano+CHL")) %>%
  ggplot(aes(x = turn.angle, fill = Chem))+
  stat_bin_(data = AGG_Dark,  position = "stack", binwidth = 15, colour = "black", size = .25)+
  coord_polar(theta = "x", start = 3.142, direction=1)+
  #geom_vline(data = Data2, aes(xintercept = ta), size = 1)+
  theme_classic()+
  scale_x_continuous(limits = c(-181,180), ## if -180 labels overlap and look funny
                     breaks = seq(-180, 180, by = 45),
                     minor_breaks = seq(-180, 180, by = 15))+
  theme(text=element_text(size=20))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  theme(legend.key = element_rect(fill = "white"))+
  labs(x = "Turn Angle (degrees)", y = "Frequency of Turn Angle", title = "")+
  scale_fill_manual(values=c("gray92", "gray", "gray35"),name = "Chemical Cue",
                    labels=c("Ambient", "Guano", "Guano+CHL"))+
 facet_grid(facets = Chem ~ Flow.rate, labeller = label_both)

###############################################################################
####     coefficient of dispersion plot
##Based on Range = (X max – X min) ⁄ (X max + X min)
range(Feeding$turn.angle)
Data2 <- aggregate(Feeding$turn.anglexysmooth, by = list(Feeding$Chem, Feeding$Flow.rate), max)
Data3 <- aggregate(Feeding$turn.anglexysmooth, by = list(Feeding$Chem, Feeding$Flow.rate), min)
CD <- Data3
CD$cd <- ((Data2$x-Data3$x)/(Data2$x+Data3$x))
str(CD)
CD <- read.csv(file = "~/Bigelow/Manuscripts/Feeding MS/CD.csv", sep = ',', header = TRUE)

CD %>%
  filter(Chem %in% c("Ambient", "Guano", "Guano+CHL")) %>%
  ggplot(aes(x = Data2$ta, y = 1))+
  coord_polar(theta = "x", start = 3.142, direction=1)+
  #geom_vline(data = Data2, aes(xintercept = ta), size = 1)+
  geom_segment(aes(x=0, y=0, xend=Data2$ta, yend=cd+.4), arrow = arrow(length=unit(.5, 'cm'), type = "open"))+
  theme_classic()+
  scale_x_continuous(limits = c(-181,180), ## if -180 labels overlap and look funny
                     breaks = seq(-180, 180, by = 45),
                     minor_breaks = seq(-180, 180, by = 15))+
  theme(text=element_text(size=20))+
  theme(axis.text=element_text(color="black"))+
  theme(axis.text.x=element_text(angle=0,hjust=1,vjust=0.4))+
  theme(axis.line = element_line(size = 0.5, linetype = "solid",
                                   colour = "black"))+
  labs(x = "Turn Angle (degrees)", y = "Coefficient of Dispersion (r)", title = "") +
  facet_grid(facets = Chem ~ Flow.rate, labeller = label_both)

########################
  
```

Circular stats for pathwise means of angular data
```{r}
library(tidyverse)
library(circular)
library(bbmle)

dvm <- circular:::DvonmisesRad
## check vectorization
stopifnot(all.equal(c(dvm(1,1,0.5,log=TRUE), dvm(2,3,0.5,log=TRUE)),
                    dvm(c(1,2), c(1,3), 0.5, log=TRUE)))

m_null <- mle2(turn.angle.smooth.degrees ~ dvm(exp(log_mu), exp(log_kappa)),
           parameters = list(log_mu ~ 1),
           data = AGG_Dark,
           start = list(log_mu = 0, log_kappa = 0))

m3_FC <- update(m_null, parameters = list(log_mu ~ Flow.Rate+Chemical.Cue))
m3_F  <- update(m_null, parameters = list(log_mu ~ Flow.Rate))


anova(m3_FC, m3_F, m_null)

```

Circular stats for angles
```{r}
### circular stats for angles
range(LML$heading.pi)
library(circular)

range(AGG_Dark$heading.smooth.degrees)
AGG_Feed$turn.anglexysmooth <- AGG_Feed$turn.anglexysmooth/(2*pi)*360

circ.turn <- circular(AGG_Dark$heading.smooth.degrees, type = c("angles"),
          units = c("degrees"),
          template = c("geographics"))

## S3 method for class 'circular'
## as(circ, control.circular=list(), ...) ## NOT WORKING CURRENTLY

## S3 method for class 'circular'
is(circ)

## S3 method for class 'circular'
print(circ, info=TRUE)
range(circ)

## S3 method for class 'circular'
plot(circ, pch = 16, cex = 1, stack = TRUE,
axes = TRUE, start.sep=0, sep = 0.025, shrink = 3.5,   ### larger stack number = more zoomed out
bins = 120, ticks = FALSE, tcl = 0.025, tcl.text = 0.125,
col = NULL, tol = 0.04, uin = NULL,
xlim = c(-1, 1), ylim = c(-1, 1), digits = 2, units = NULL,
template = NULL, zero = NULL, rotation = NULL,
main = NULL, sub=NULL, xlab = "", ylab = "",
control.circle=circle.control())


### calculating mean vector from circular data
mean(circ.turn)

## calculating variance of the vector from circular data
var(circ.turn)

## calculating mean deviation from circular data
meandeviation(circ.turn)

# Compute summary statistics of a random sample of observations. 
summary(circ.turn) 

## ANOVA using circular stats
aov.circular(circ.turn, AGG_Dark$Chemical.Cue) ## working but unsure

```

Rayleigh and Watson Two Test
```{r}

##  A Rayleigh Test is a test for significant unimodal orientation (e.g. Is everyone in the treatment going the same direction?).
head(AGG_Dark)
AGG_DM <- subset(AGG_Dark, Flow.Rate == "Medium Flow")
AGG_DH <- subset(AGG_Dark, Flow.Rate == "High Flow")

circDH <- circular(AGG_DH$heading.smooth.degrees, type = c("angles"),
          units = c("degrees"),
          template = c("geographics"))

rayleigh.test(circDM)  
rayleigh.test(circDH) 

# Test whether LNN and LEN (etc) orient differently  

##  A Watson Test determines if two groups’ orientations are significantly different from each other. 
watson.two.test(circDM, circDH, alpha = 0.05)

```

Tests against each other
```{r}   
################################################### Example of testing means of groups against each other
# Ant orientation from Duelli and Wehner (1973) 
# Example used in Batschelet (1981) 


data1 <- list(dahi = circular(AGG_DH$heading.smooth.degrees, units="degrees", template="geographics"), dame = circular(AGG_DM$heading.smooth.degrees, units="degrees", template="geographics"))
           
watson.williams.test(data1)

## gives error 
  ##Warning in watson.williams.test.default(x, group) :
  ##Concentration parameters (0.328, 0.231) not equal between groups. The test might not be applicable
##Warning in watson.williams.test.default(x, group) :
  ##Global concentration parameter: 0.314 < 2. The test is probably not applicable

    ## suggests using Wheeler-Watson test instead

watson.wheeler.test(data1)

## gives warning of :
## Warning in watson.wheeler.test.default(x, group) :
  ##There are 75884 ties in the data.
  ##Ties will be broken appart randomly and may influence the result.
  ##Re-run the test several times to check the influence of ties.

          ## after running 4 times: W = (792.65, .78, .8, .7)
          ## all df's = 2
          ## all p-values = 2.2

```

Binning Headings into 15 degree groups then doing a gtest
```{r}
library(AMR)
range(AGG_Dark$heading.smooth.degrees)
## need to bin into 15 degree bins
AGG_Dark$head.bins <- cut(AGG_Dark$heading.smooth.degrees, breaks=c(0,15,30,45,60,75,90,105,120,135,150,165,180,195,210,225,240,255,270,285,300,315,330,345,360), labels=c("0-15","15-30","30-45", "45-60", "60-75", "75-90", "90-105", "105-120", "120-135", "135-150", "150-165", "165-180", "180-195", "195-210", "210-225", "225-240", "240-255", "255-270", "270-285", "285-300", "300-315", "315-330", "330-345", "345-360"))
str(AGG_Dark)
g.test(AGG_Dark$Flow.Rate, AGG_Dark$head.bins)

write.csv(FeedingDark, file = "~/Bigelow/Manuscripts/Feeding MS/FeedingDark.csv", row.names = FALSE)
write.csv(AGG_Dark, file = "~/Bigelow/Manuscripts/Feeding MS/AGG_Dark.csv", row.names = FALSE)
```
